% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{ageRich}
\alias{ageRich}
\title{Counts species richness at a point}
\usage{
ageRich(dat)
}
\arguments{
\item{dat}{}
}
\value{
dataframe
}
\description{
Selects model from tabled models per species
}
