% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{air_density}
\alias{air_density}
\title{Finds air density}
\usage{
air_density(P, T_K, Rgas = 287.058)
}
\arguments{
\item{P}{Atmospheric pressure}

\item{T_K}{Temperature (K)}

\item{Rgas}{Specific gas constant for dry air}
}
\value{
Value
}
\description{
Finds air density
}
