% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{ausTraitTable}
\alias{ausTraitTable}
\title{ausTraitTable}
\usage{
ausTraitTable(version = "3.0.2", path = "data/austraits", shadeTolerance = 3)
}
\arguments{
\item{version}{Version of austraits}

\item{path}{Path to austraits database}

\item{shadeTolerance}{The Whole Plant Light Compensation Point below which 
plants are considered shade tolerant}
}
\value{
list of tables
}
\description{
Constructs a default.species.params table using traits available in the austraits database
Collects traits from ausTraits for use in shade and fire effects modelling
}
