% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{bole}
\alias{bole}
\title{bole}
\usage{
bole(
  lengthSurface = 2,
  residence = 300,
  depth = 0.05,
  h = 0.1,
  surfDecl = 10,
  t = 1
)
}
\arguments{
\item{lengthSurface}{Length of the surface flame (m)}

\item{residence}{Residence time of the flame (s)}

\item{depth}{Depth of the surface litter (m)}

\item{h}{Height of the bole above ground (m)}

\item{surfDecl}{Rate of decline of the surface flame length (m/s)}

\item{t}{Time since the flame front passed (s)}
}
\value{
value
}
\description{
Air temperature above ambient at the tree bole behind the flame front

Dynamic air temperature at bole height, declining flame length exponentially

Air temperature is modelled from dynamic flame segments using
Weber R.O., Gill A.M., Lyons P.R.A., Moore P.H.R., Bradstock R.A., Mercer G.N. (1995)
Modelling wildland fire temperatures. CALMScience Supplement, 4, 23–26.

pAlphas is set to bole height - depth of surface litter
}
