% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{buildStructure}
\alias{buildStructure}
\title{buildStructure}
\usage{
buildStructure(
  veg,
  overlap = 0.5,
  sepSig = 0.001,
  pN = "Point",
  spName = "spName",
  base = "base",
  top = "top",
  he = "he",
  ht = "ht"
)
}
\arguments{
\item{veg}{A dataframe listing plant species with columns describing crown dimensions using standardised names:
veg, pN, spName, base, top, he, ht, wid, Site, sN}

\item{overlap}{Either 'automatic', or threshold occurrence at which overlap is set to TRUE.}

\item{sepSig}{Significance at which to recognise separate strata}

\item{pN}{Name of the column designating the point identifier}

\item{spName}{Name of the column designating the species name}

\item{base}{Name of the column with the canopy base height}

\item{top}{Name of the column with the canopy top height}

\item{he}{Name of the column with the he height}

\item{ht}{Name of the column with the ht height}
}
\value{
dataframe
}
\description{
Constructs the table F_structure from formatted survey data
}
