% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{buildStructurePP}
\alias{buildStructurePP}
\title{Creates an 'F_structure' table for FRaME from plant modelling
Discontinued}
\usage{
buildStructurePP(dat, age, rec = 1)
}
\arguments{
\item{dat}{The output from stratify_community}

\item{age}{Years since disturbance}

\item{rec}{Number of the record}
}
\description{
Creates an 'F_structure' table for FRaME from plant modelling
Discontinued
}
