% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{buildTraitsP}
\alias{buildTraitsP}
\title{buildTraitsP}
\usage{
buildTraitsP(
  comm,
  propDead = 0,
  leafForm = "Flat",
  lwRat = 3,
  leafA = 0.002547,
  ram = 5,
  ignitionTemp = 260,
  moist = 1,
  G.C_rat = 3,
  C.C_rat = 0.1,
  deltaL = 0.46
)
}
\arguments{
\item{comm}{The output from stratify_community}

\item{propDead}{Proportion of foliage dead}

\item{leafForm}{Flat or Round}

\item{lwRat}{Ratio of leaf length to width}

\item{leafA}{Area of a leaf in m2}

\item{ram}{Stem ramification}

\item{ignitionTemp}{Temperature of the endotherm (degC)}

\item{moist}{Leaf moisture (ratio moisture weight to dry weight)}

\item{G.C_rat}{Ratio of gaps to clumps of leaves}

\item{C.C_rat}{Ratio of clump to canopy size}

\item{deltaL}{Leaf density (g/cm3)}
}
\description{
Creates traits table for FRaME from plant modelling

Leaf dimensions default to ausTraits mean leaf size
}
