% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{canopyCheck}
\alias{canopyCheck}
\title{Modifies inputs ensuring params file fits criteria}
\usage{
canopyCheck(
  base.params,
  testN = 5,
  Flora,
  Structure,
  a = 1,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  threshold = 1
)
}
\arguments{
\item{base.params}{}

\item{testN}{Number of replicates to allow}

\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100% ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - % composition or count of that species in the stratum. If absent, all species will be considered equally
hc, he, ht, hp & w - canopy dimensions for that species (m)
clump - mean ratio of clump diameter to crown diameter
openness - proportion of plant canopy occupied by gaps between clumps}

\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. Acceptable values
are t, f, or blank, where the outcome will be decided by the relative stratum heights.}

\item{a}{A unique identifier for the record being run}

\item{l}{}

\item{Ms}{}

\item{Pm}{}

\item{Mr}{}
}
\value{
Dataframe
}
\description{
Modifies inputs ensuring params file fits criteria
}
