% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMC.R
\name{climDynamics}
\alias{climDynamics}
\title{climDynamics}
\usage{
climDynamics(
  fireDat,
  clim,
  m = 0.15,
  slope = 0,
  slopeSD = 0,
  rholitter = 550,
  lat = -34.95,
  lAngleAccounting = TRUE,
  rangeDir = 270,
  cardinal = TRUE,
  freeCores = 1
)
}
\arguments{
\item{fireDat}{A list with three dataframes: Flora, Structure, and default.species.params.}

\item{clim}{A dataset with the fields:
tAM (9am temp, degC)
tPM (3pm temp, degC)
tMin (Minimum daily temp, degC)
tMax (Maximum daily temp, degC)
rhAM (9am Relative humidity, percent)
rhPM (3pm Relative humidity, percent)
wAM (9am wind, km/h)
wPM (3pm wind, km/h)
cAM (Morning cloud, oktas)
cPM (Afternoon cloud, oktas)
MSLP (Mean Sea Level Pressure, hPa)
Rain (Daily rainfall, mm)}

\item{m}{The starting moisture (proportion ODW)}

\item{slope}{(degrees)}

\item{slopeSD}{Standard deviation of the slope}

\item{rholitter}{Density of litter particles (kg/m3)}

\item{lat}{Latitude (degrees)}

\item{lAngleAccounting}{Set to TRUE to automatically modify leaf angle by number of strata}

\item{rangeDir}{Cardinal direction of the ridgelines - either Nth/Sth (0) or west/east (270)}

\item{cardinal}{Simplify compass directions to cardinal (TRUE/FALSE)}

\item{freeCores}{Number of cores to leave unused by the model}
}
\value{
dataframe
}
\description{
Models input weather parameters from a climate dataset,
each age is modelled on a separate core
}
