% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_params_functions.R
\name{completeParams}
\alias{completeParams}
\title{Completes a table by adding default parameters for each species as required.}
\usage{
completeParams(tbl)
}
\arguments{
\item{tbl}{The input parameter table}
}
\value{
The completed parameter table
}
\description{
For each species in the input parameter table, this function checks which
parameters have been provided and retrieves defaults for missing parameters
from the default lookup table: \link{DefaultSpeciesParams}.
}
\details{
If a species has missing parameters but is not present in the table of
default lookup table, an error is thrown.
}

