% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{conDrivers}
\alias{conDrivers}
\title{Models fire behaviour across defined slopes and DFMCs, 
with varied plants and random winds and within a defined range}
\usage{
conDrivers(
  base.params,
  db.path = "out_mc.db",
  replicates,
  windMin,
  windMax,
  slopes = c(0, 10),
  DFMCs = c(0.05, 0.1, 0.15),
  moistureMultiplier,
  moistureSD,
  moistureRange,
  heightSD,
  heightRange,
  leafVar,
  updateProgress = NULL
)
}
\arguments{
\item{base.params}{Input parameter file}

\item{db.path}{Name of the exported database}

\item{replicates}{Number of wind replicates per slope/DFMC combination}

\item{windMin}{Lowest wind velocity tested (km/h)}

\item{windMax}{Maximum wind velocity tested (km/h)}

\item{slopes}{A string of slope values to be tested}

\item{DFMCs}{A string of dead fuel moisture values to be tested}

\item{moistureMultiplier}{Multiplies all LFMC values by this number}

\item{moistureSD}{Standard deviation of moisture}

\item{moistureRange}{Truncates variability by +/- mean * range}

\item{heightSD}{Standard deviation of plant height}

\item{heightRange}{Truncates variability by +/- mean * range}

\item{leafVar}{Variation around input leaf dimensions, equivalent to l}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\description{
Private function under development
}
