% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{coverDyn}
\alias{coverDyn}
\title{coverDyn}
\usage{
coverDyn(
  dat,
  species_col = "Species",
  grouping = "Age",
  thres = 5,
  pN = "Point",
  pntGap = 5,
  p = 0.05,
  bTest = 10,
  maxiter = 1000
)
}
\arguments{
\item{dat}{The dataframe containing the input data,}

\item{species_col}{The name of the column giving the name of the species (string)}

\item{grouping}{The name of the column that groups sites, eg by age or location (string)}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{pN}{The name of the column giving the number of the vertical transect (string)}

\item{pntGap}{The distance (m) between each vertical transect point}

\item{p}{The maximum allowable p value for a model}

\item{bTest}{Multiples of mean + mRSE for which Burr & quadratic models can predict
beyond the observed mean + standard deviation}

\item{maxiter}{The maximum number of iterations for model fitting}
}
\value{
dataframe
}
\description{
Builds models for cover dynamics of surveyed Species
Species that are less common than the set threshold are combined as "Minor Species"
}
