% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threat.R
\name{cp}
\alias{cp}
\title{cp}
\usage{
cp(Material = "wood", temp, moist)
}
\arguments{
\item{Material}{Type of material, either 'bark' or 'wood'}

\item{temp}{Temperature (C)}

\item{moist}{Moisture content (proportion oven dry weight)}
}
\value{
value
}
\description{
Specific heat capacity of a material
}
