% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earth.R
\name{cpSoil}
\alias{cpSoil}
\title{cpSoil}
\usage{
cpSoil(temp = 300, texture = "loam", peat = 0.2, moisture = 0.3)
}
\arguments{
\item{temp}{Temperature of the soil (K)}

\item{texture}{Soil texture}

\item{peat}{Organic proportion of the soil}

\item{moisture}{Volumetric water content}
}
\description{
Finds specific heat of soil

Finds volumetric specific heat from the mineral, organic and water components of the soil

Specific heats of soil components estimated from figures 108 & 111 in
Farouki, O. Thermal properties of soils. (Trans Tech, 1981).

Water specific heat 4185 J/kg.K
}
