% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DefaultSpeciesParams}
\alias{DefaultSpeciesParams}
\title{Species trait database.}
\format{A data frame where each row is a species; species name in
  the first column; trait values in subsequent columns.}
\usage{
DefaultSpeciesParams
}
\description{
Provides default parameter values for commonly referenced species.
}
\keyword{datasets}

