% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DefaultUnits}
\alias{DefaultUnits}
\title{Default units for model parameters.}
\format{A data frame with columns: section (values meta,
stratum, species); param (standard parameter label);
units (unit abbreviation or \code{NA} if dimensionless).}
\usage{
DefaultUnits
}
\description{
These units apply when none are specified by the user.
}
\details{
A value of \code{NA} indicates that a parameter is dimensionless (e.g. a proportional 
quantity or a text value).
}
\keyword{datasets}

