% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLegacyParamFile.R
\name{.parse_params}
\alias{.parse_params}
\title{parse_params}
\usage{
.parse_params(
  dat,
  in.param.col = "txt",
  in.delim.col = "delim",
  out.name.col = "param",
  out.value.col = "value"
)
}
\arguments{
\item{dat}{tbl or dbf object}

\item{in.param.col}{input 1}

\item{in.delim.col}{input 2}

\item{out.name.col}{output 1}

\item{out.value.col}{output 2}
}
\value{
A data frame with parsed parameters.
}
\description{
parse_params
}
