% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLegacyParamFile.R
\name{.parse_text}
\alias{.parse_text}
\title{parse_text}
\usage{
.parse_text(txt)
}
\arguments{
\item{txt}{A character vector of text lines.}
}
\description{
Parses pre-processed parameter file text
into a structured form.
}
