% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{drivers}
\alias{drivers}
\title{Models fire behaviour across ranged variables}
\usage{
drivers(
  base.params,
  db.path = "out_mc.db",
  jitters,
  windMin,
  windReps,
  windStep,
  moistureMultiplier,
  moistureSD,
  moistureRange,
  heightSD,
  heightRange,
  leafVar,
  updateProgress = NULL
)
}
\arguments{
\item{base.params}{Input parameter file}

\item{db.path}{Name of the exported database}

\item{jitters}{Number of repetitions for each row in the weather table}

\item{windMin}{Lowest wind velocity tested (km/h)}

\item{windReps}{Number of wind speeds tested}

\item{windStep}{Gap (km/h) between wind steps}

\item{moistureMultiplier}{Multiplies all LFMC values by this number}

\item{moistureSD}{Standard deviation of moisture}

\item{moistureRange}{Truncates variability by +/- mean * range}

\item{heightSD}{Standard deviation of plant height}

\item{heightRange}{Truncates variability by +/- mean * range}

\item{leafVar}{Variation around input leaf dimensions, equivalent to l}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\description{
Models fire behaviour across ranged variables
}
\examples{

SPECIFY INPUTS
record <- 1
data(site)
data(structure)
data(flora)
data(traits)
base.params <- paramBuilder(site, structure, flora, traits, record)

MODEL PROBABILISTIC FIRE BEHAVIOUR
drivers(base.params, db.path = "out.db", jitters = REPLICATES, windMin = 0, windReps = 30,
windStep = 2, moistureMultiplier = 1, moistureSD = 0.01, moistureRange = 1.001, 
heightSD = 0.2, heightRange = 1.41, leafVar = 0.1, updateProgress = NULL)

LOAD AND ORGANISE RESULTS
res<-ffm_db_load("out.db")
}
