% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{driversFrame}
\alias{driversFrame}
\title{Models fire behaviour across ranged variables using species specific details}
\usage{
driversFrame(
  Flora,
  Structure,
  default.species.params,
  a,
  db.path = "out_mc.db",
  replicates = 3,
  winds = seq(0, 60, 5),
  slopes,
  DFMCs,
  tAir,
  moistureMultiplier,
  moistureSD,
  moistureRange,
  leafVar,
  testN = 10,
  updateProgress = NULL
)
}
\arguments{
\item{Flora}{}

\item{Structure}{}

\item{default.species.params}{}

\item{a}{A unique identifier for the record being run}

\item{db.path}{Name of the exported database}

\item{replicates}{}

\item{winds}{}

\item{slopes}{List of slope values for testing}

\item{DFMCs}{List of DFMC values for testing}

\item{tAir}{}

\item{moistureMultiplier}{Multiplies all LFMC values by this number}

\item{moistureSD}{Standard deviation of moisture}

\item{moistureRange}{Truncates variability by +/- mean * range}

\item{leafVar}{Variation around input leaf dimensions, equivalent to l}

\item{testN}{}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\description{
Models fire behaviour across ranged variables using species specific details
}
