% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{driversRand}
\alias{driversRand}
\title{Models fire behaviour across ranged variables using species specific details}
\usage{
driversRand(
  base.params,
  a,
  db.path = "out_mc.db",
  replicates,
  windMin,
  windReps,
  windStep,
  slopes,
  DFMCs,
  moistureMultiplier,
  moistureSD,
  moistureRange,
  temperature,
  Variation,
  leafVar,
  updateProgress = NULL
)
}
\arguments{
\item{base.params}{Input parameter file}

\item{a}{A unique identifier for the record being run}

\item{db.path}{Name of the exported database}

\item{windMin}{Lowest wind velocity tested (km/h)}

\item{windReps}{Number of wind speeds tested}

\item{windStep}{Gap (km/h) between wind steps}

\item{slopes}{List of slope values for testing}

\item{DFMCs}{List of DFMC values for testing}

\item{moistureMultiplier}{Multiplies all LFMC values by this number}

\item{moistureSD}{Standard deviation of moisture}

\item{moistureRange}{Truncates variability by +/- mean * range}

\item{temperature}{Standardised air temperature for the test}

\item{Variation}{A database of plant variability in traits, with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
stratum - numeric value from 1 to 4, counting from lowest stratum
Hs - Standard deviation of plant height variations
Hr - Truncates plant height variability by +/- Hr * height}

\item{leafVar}{Variation around input leaf dimensions, equivalent to l}

\item{updateProgress}{Progress bar for use in the dashboard}

\item{jitters}{Number of repetitions for each row in the weather table}
}
\description{
Models fire behaviour across ranged variables using species specific details
}
