% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{entrainment_depth_angle}
\alias{entrainment_depth_angle}
\title{Finds entrainment depth at a given angle}
\usage{
entrainment_depth_angle(
  Plume_velocity,
  lengthHollow,
  beta_deg = 90,
  Cp_cross = 0.9,
  Cp_stag = 1,
  Altitude = 0,
  hollowTemp = 20,
  tempAir = 100,
  mode = c("adiabatic", "isothermal"),
  gamma = 1.4,
  Pressure = 101325,
  ambientTemp = 281.65,
  lapseRate = 0.0065,
  g = 9.80665,
  Rgas = 287.058
)
}
\arguments{
\item{Plume_velocity}{Velocity of airflow (m/s)}

\item{lengthHollow}{Length of the hollow (m)}

\item{beta_deg}{Angle of the hollow to the airflow (degrees). 0 = facing the flow, 90 = perpendicular to the flow}

\item{Cp_cross}{Pressure coefficient for suction in crossflow (typically 0.4–1.0)}

\item{Cp_stag}{Pressure coefficient for stagnation at head-on flow}

\item{Altitude}{Height above sea level (m)}

\item{hollowTemp}{Starting temperature inside the hollow (C)}

\item{tempAir}{Temperature of the convective plume heating the hollow (C)}

\item{mode}{Considers expansion law for trapped gas}

\item{gamma}{Ratio of specific heats for air}

\item{Pressure}{Atmospheric pressure at sea level (Pa)}

\item{ambientTemp}{Ambient temperature at the site (K)}

\item{lapseRate}{K/m (standard tropospheric lapse rate)}

\item{g}{Acceleration due to gravity (m/s^2)}

\item{Rgas}{Specific gas constant for dry air J/(kg·K)}
}
\value{
Value
}
\description{
Finds entrainment depth at a given angle
}
