% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{entrainmentHeating}
\alias{entrainmentHeating}
\title{Heating of a hollow through air entrainment}
\usage{
entrainmentHeating(
  Plume_velocity,
  lengthHollow,
  diameterHollow,
  Altitude = 0,
  hollowTemp = 30,
  tempAir = 100,
  Pressure = 101325,
  ambientTemp = 281.65,
  times = seq(0, 60, by = 1),
  beta_deg = 90,
  Cp_cross = 0.9,
  Cp_stag = 1,
  mode = c("adiabatic", "isothermal"),
  D_mol = 2e-05,
  C_t = 0.02,
  k = 2,
  lapseRate = 0.0065,
  g = 9.80665,
  Rgas = 287.058
)
}
\arguments{
\item{Plume_velocity}{Velocity of airflow (m/s)}

\item{lengthHollow}{Length of the hollow (m)}

\item{diameterHollow}{Diameter of the hollow (m)}

\item{Altitude}{Height above sea level (m)}

\item{hollowTemp}{Starting temperature inside the hollow (C)}

\item{tempAir}{Temperature of the convective plume heating the hollow (C)}

\item{Pressure}{Atmospheric pressure at sea level (Pa)}

\item{ambientTemp}{Ambient temperature at the site (K)}

\item{times}{Time sequence to evaluate (s)}

\item{beta_deg}{Angle of the hollow to the airflow (degrees). 0 = facing the flow, 90 = perpendicular to the flow}

\item{Cp_cross}{Pressure coefficient for suction in crossflow (typically 0.4–1.0)}

\item{Cp_stag}{Pressure coefficient for stagnation at head-on flow}

\item{mode}{Considers expansion law for trapped gas}

\item{D_mol}{Molecular diffusivity of air (m^2/s)}

\item{C_t}{Coefficient for shear-enhanced turbulent diffusivity}

\item{k}{Coefficient for depth of mixing layer in the hollow}

\item{lapseRate}{K/m (standard tropospheric lapse rate)}

\item{g}{Acceleration due to gravity (m/s^2)}

\item{Rgas}{Specific gas constant for dry air J/(kg·K)}
}
\value{
value
}
\description{
Heating of a hollow through air entrainment
}
