% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f_site}
\alias{f_site}
\title{Site data}
\format{
a dataframe with 8 columns, and a row for each site
\describe{
\item{record}{Consecutive integer to reference the row}
\item{site}{A string to name the row}
\item{slope}{Slope (degrees)}
\item{wind}{Wind velocity (km/h)}
\item{temp}{Ambient air temperature (deg.C)}
\item{dfmc}{Dead fuel moisture content (proportion oven dry weight)}
\item{litter}{Weight (t/ha) of fine surface litter}
\item{fLine}{Length of the active fire line (m)}
\item{diameter}{mean diameter of litter particles (m)}
}
}
\usage{
f_site
}
\description{
A data frame containing the site data for the model
}
\keyword{datasets}
