% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f_traits}
\alias{f_traits}
\title{Plant trait data}
\format{
a data frame with 9 columns, and a row for each species
\describe{
\item{name}{The name of the species}
\item{propDead}{Proportion of the foliage that is dead}
\item{leafForm}{Allowable values are flat or round (e.g. terete)}
\item{leafThickness}{Thickness of the leaf (m)}
\item{leafWidth}{Width at the widest axis of the leaf (m)}
\item{leafLength}{Length of the leaf (m)}
\item{leafSeparation}{Separation between leaves along a stem}
\item{stemOrder}{Stem ramification}
\item{ignitionTemp}{Piloted ignition temperature of the leaf (deg.C)}
\item{Field}{Description}
}
}
\usage{
f_traits
}
\description{
A data frame containing the plant trait data for the model
}
\keyword{datasets}
