% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f_weather}
\alias{f_weather}
\title{Weather data}
\format{
a data frame with 4 columns and a row for each time step
\describe{
\item{tm}{Consecutive integer to reference the time period}
\item{T}{Ambient air temperature (deg. C)}
\item{W}{Wind velocity (km/h)}
\item{DFMC}{Dead fuel moisture content (proportion oven dry weight)}

}
}
\usage{
f_weather
}
\description{
A data frame containing the weather data for the model
}
\keyword{datasets}
