% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_cleanup}
\alias{ffm_cleanup}
\title{Attempts to free up resources used for R/Scala connection.}
\usage{
ffm_cleanup(reset = FALSE)
}
\arguments{
\item{reset}{Whether to also re-start the Scala interpreter used
by the package (default is \code{FALSE}).}
}
\description{
This is generally for use by other package functions, but can
also be run by the user. To be effective, all references to 
Scala objects should be removed from the workspace first.
}
\note{
This function is deprecated and will be removed in a future
  version of the package
}
