% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_db_functions.R
\name{ffm_db_connect}
\alias{ffm_db_connect}
\title{Opens a connection to a SQLite database of model results.}
\usage{
ffm_db_connect(db.path)
}
\arguments{
\item{db.path}{The path and filename for the database.}
}
\value{
A connection object.
}
\description{
This function attempts to connect to a SQLite database
file. If successful it returns a connection which can be
used with functions in the RSQLite or dplyr packages to query data.
}
\examples{
\dontrun{
# Querying the database directly with RSQLite functions

library(RSQLite)

# Open a connection
con <- ffm_db_connect("c:/michael/somewhere/my_results.db")

# Query the FlameSummaries table for maximum flame height
# in each run
sql <- "SELECT repId, MAX(flameHeight) AS height FROM FlameSummaries GROUP BY repId"
res <- dbGetQuery(con, sql)

# Close the connection when you are finished with it
dbDisconnect(con)

# TODO - An example of querying the database with
# the dplyr package
}

}

