% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_db_functions.R
\name{ffm_db_summary}
\alias{ffm_db_summary}
\title{ffm_db_summary}
\usage{
ffm_db_summary(db.path)
}
\arguments{
\item{db.path}{Path and filename for the database.}
}
\description{
Summarize the contents of a model results database.

At the moment this function simply prints the number of
simulation runs found in the database.
}
