% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_functions.R
\name{ffm_describe_betas}
\alias{ffm_describe_betas}
\title{Generates example Beta distributions.}
\usage{
ffm_describe_betas(
  beta.a = c(2, 5, 10),
  beta.b = c(2, 5, 10),
  show.plot = TRUE
)
}
\arguments{
\item{beta.a}{Vector of one or more values for parameter \code{a}
(called \code{shape1} by R beta functions).}

\item{beta.b}{Vector of one or more values for parameter \code{b}
(called \code{shape2} by R beta functions).}

\item{show.plot}{If \code{TRUE}, display the resulting distributions.}
}
\value{
Invisibly returns a list with elements:
\describe{
\item{plot}{A ggplot object showing the distributions.}
\item{summary}{A data frame summarizing the moments of the distributions.}
}
}
\description{
This function can be used to explore the shapes and parameters of
alternative Beta distributions. Given one or more values for
each of the two Beta parameters, it calculates the density
curve for each combination plus the mean, mode and variance of
the distribution.
}
\details{
Note that this function will draw and describe Beta distributions
with any valid parameter values, but the \code{\link{ffm_param_variance}}
function only accepts distributions where both parameters are greater
than one to ensure that the mode exists.
}
\seealso{
\code{\link{ffm_param_variance}} which has an option to use
a Beta distribution to vary parameter values.
}
