% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_functions.R
\name{ffm_disassemble_table}
\alias{ffm_disassemble_table}
\title{Disassembles a parameter table into a list of components.}
\usage{
ffm_disassemble_table(tbl)
}
\arguments{
\item{tbl}{The input simulation parameter table (data frame).}
}
\value{
A named list of table components.
}
\description{
Given a parameter table, this function returns a list with
the following named elements:
\describe{
\item{site.meta}{Data frame of rows from the input table for
site attributes.}
\item{strata.meta}{Data frame of rows from the input table for
stratum meta-data.}
\item{species.values}{Data frame of species parameter values in
wide format, ie. species as rows, parameters as columns.}
\item{species.units}{Data frame of species parameter units in
wide format.}
}
For the \code{species.values} data frame, all columns for numeric
parameters are converted from character to numeric.
}
\seealso{
\code{\link{ffm_assemble_table}} for the reverse process.
}
