% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_find_site_param}
\alias{ffm_find_site_param}
\title{Locate specified site meta-parameters in a parameter table.}
\usage{
ffm_find_site_param(params, names)
}
\arguments{
\item{params}{Parameter table as a matrix or data.frame of character values.}

\item{names}{The name(s) of parameters to find.}
}
\value{
Row numbers corresponding to the specified parameters.
}
\description{
Locate specified site meta-parameters in a parameter table.
}

