% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_functions.R
\name{ffm_get_species_params}
\alias{ffm_get_species_params}
\title{Gets default parameter values for one or more species.}
\usage{
ffm_get_species_params(names)
}
\arguments{
\item{names}{The names, whole or partial, to search for.}
}
\value{
A data frame with default parameter values for each matched species.
}
\description{
Takes a character vector of one or more species names (whole or partial)
and retrieves the default parameter values for each. If one or more
names is not found, or matches mutliple species in the default
parameters table, an error message is given.
}
\examples{
\dontrun{
# Get default parameters for selected species:
spp <- c("Poa lab", "D repens", "Hyd laxiflora")
params <- ffm_get_species_params(spp)
}

}

