% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_functions.R
\name{ffm_param_info}
\alias{ffm_param_info}
\title{Retrieves information for a parameter label.}
\usage{
ffm_param_info(label, section = NULL, no.match.error = FALSE)
}
\arguments{
\item{label}{The parameter label.}

\item{section}{If not \code{NULL}, one of site, stratum or species.}

\item{no.match.error}{If \code{TRUE} an error results when a unique
match is not found for the input label; if \code{FALSE} (default)
the function returns NULL when there is no unique match.}
}
\value{
Parameter info as a single row data frame; or \code{NULL}
if no matching label was found and \code{no.match.error} is
\code{FALSE}.
}
\description{
Searches for a row in the \code{\link{ParamInfo}} table with a matching
parameter label. The comparison ignores case and white-space.
If a non-NULL value is supplied for \code{section} the check is restricted
to parameter labels in that section (site, stratum or species).
}
