% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run.data.frame}
\alias{ffm_run.data.frame}
\title{Run the model with a site based on parameters in the
given data frame.}
\usage{
\method{ffm_run}{data.frame}(site.params, db, ...)
}
\arguments{
\item{site.params}{A data frame of site parameters valid for
input to \code{\link{ffm_site}}}
}
\description{
Run the model with a site based on parameters in the
given data frame.
}
