% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run.function}
\alias{ffm_run.function}
\title{Run the model with parameters provided by a generating function.}
\usage{
\method{ffm_run}{function}(gen.fn, db, ...)
}
\arguments{
\item{gen.fn}{A function which generates a valid parameter matrix.}
}
\description{
Run the model with parameters provided by a generating function.
}

