% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run.matrix}
\alias{ffm_run.matrix}
\title{Run the model with a site based on parameters in the
given character matrix.}
\usage{
\method{ffm_run}{matrix}(site.params, db, ...)
}
\arguments{
\item{site.params}{A character matrix of site parameters valid for
input to \code{\link{ffm_site}}}
}
\description{
Run the model with a site based on parameters in the
given character matrix.
}

