% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run.ScalaCachedReference}
\alias{ffm_run.ScalaCachedReference}
\title{Run the model with the given site object.}
\usage{
\method{ffm_run}{ScalaCachedReference}(site, db, ...)
}
\arguments{
\item{site}{The site as a Scala reference object created
with \code{\link{ffm_site}}.}

\item{db}{Output database object or path and filename.}
}
\value{
\code{TRUE} if the run completed and results were written to
  the output database successfully; \code{FALSE} otherwise.
}
\description{
Run the model with the given site object.
}
\seealso{
\code{\link{ffm_create_database}}
}
