% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run.ScalaInterpreterReference}
\alias{ffm_run.ScalaInterpreterReference}
\title{Run the model with the given site object.}
\usage{
\method{ffm_run}{ScalaInterpreterReference}(site)
}
\arguments{
\item{site}{The site as a ScalaInterpreterReference object created
with \code{\link{ffm_site}}.}
}
\value{
???
}
\description{
Run the model with the given site object.
}

