% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{ffm_run_robust}
\alias{ffm_run_robust}
\title{Modifies inputs until model runs}
\usage{
ffm_run_robust(
  base.params,
  db.path,
  db.recreate = TRUE,
  testN = 5,
  Strata,
  Species,
  Flora,
  Structure,
  a = 1,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5
)
}
\arguments{
\item{base.params}{}

\item{db.path}{}

\item{db.recreate}{}

\item{testN}{Number of replicates to allow}

\item{Strata}{}

\item{Species}{}

\item{Flora}{}

\item{Structure}{}

\item{a}{}

\item{l}{}

\item{Ms}{}

\item{Pm}{}

\item{Mr}{}
}
\value{
\code{TRUE} if the run completed and results were written to
  the output database successfully; \code{FALSE} otherwise.
}
\description{
Modifies inputs until model runs
}
