% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_functions.R
\name{ffm_set_species_param}
\alias{ffm_set_species_param}
\title{Sets the value of a species parameter.}
\usage{
ffm_set_species_param(tbl, stratum.id, species.id, param, value,
  units = NA_character_)
}
\arguments{
\item{tbl}{The parameter table in which to set the value.}

\item{stratum.id}{(character) Stratum identifier.}

\item{species.id}{(character) Species identifier.}

\item{param}{(character) Label of the parameter to set.}

\item{value}{Value of the parameter (will be converted to character).}

\item{units}{(option) The units of measurement for the supplied value.}
}
\value{
The updated parameter table.
}
\description{
Sets the value of a species parameter.
}

