% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_functions.R
\name{ffm_write_params}
\alias{ffm_write_params}
\title{Writes a table of parameters to a CSV format file.}
\usage{
ffm_write_params(tbl, path, overwrite = FALSE,
  must.be.complete = FALSE)
}
\arguments{
\item{tbl}{The input simulation parameter table (data frame).}

\item{path}{The path and file name to write to. The recommended file
extension is \code{'.csv'} but this is not enforced by the function.}

\item{overwrite}{Set to \code{FALSE} (default) to issue an error message if
the output file exists, or \code{TRUE} to overwrite any existing file.}

\item{must.be.complete}{Set to \code{FALSE} (default) to enable writing a
partial table of parameters, or to \code{TRUE} to abort writing and issue an
error message if the parameter table is not complete.}
}
\description{
This is simply a wrapper that calls the standard
\code{write.csv} function with appropriate default arguments.
}
