% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_params_functions.R
\name{findSpecies}
\alias{findSpecies}
\title{Matches whole or partial species names to DefaultSpeciesParams.}
\usage{
findSpecies(names)
}
\arguments{
\item{names}{The names, whole or partial, to search for.}
}
\value{
a named \code{list} where names are the search strings and
  each value is the index or indices of matching species (NA for no match).
}
\description{
For each name in the input character vector \code{names}, finds
the index or indices of matching rows in the table of default species
parameter values.
}
\examples{
\dontrun{
# Look for Acacia dealbata and Cassinia aculeata.
# The search is case insensitive.
index <- findSpeciesByName(c("Ac deal", "cass ac"))

# Search for a term that matches multiple species. The
# list element for "cass" will be a vector of indices of
# all matching taxa:
findSpeciesByName("cass")

# Search for a name that does not exist in the default
# species parameters table. This results in \\code{NA} being
# returned for the index:
findSpeciesByName("aspidistra")
}

}
\seealso{
\code{\link{getSpeciesParams}}, \link{DefaultSpeciesParams}
}

