% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{fireDynamics}
\alias{fireDynamics}
\title{Models probabilistic fire behaviour across multiple ages on parallel cores}
\usage{
fireDynamics(
  fireDat,
  slope = 5,
  slopeSD = 2,
  slopeRange = 5,
  temp = 30,
  tempSD = 5,
  tempRange = 3,
  DFMC = 0.1,
  DFMCSD = 0.01,
  DFMCRange = 2,
  wind = 10,
  windSD = 5,
  windRange = 5,
  fLine = 1000,
  moistureMultiplier = 1,
  moistureSD = 0.01,
  moistureRange = 1.5,
  reps = 5,
  leafVar = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  vAir500 = 2,
  testN = 5,
  updateProgress = NULL,
  freeCores = 1
)
}
\arguments{
\item{fireDat}{}

\item{slope}{}

\item{slopeSD}{}

\item{slopeRange}{}

\item{temp}{}

\item{tempSD}{}

\item{tempRange}{}

\item{DFMC}{}

\item{DFMCSD}{}

\item{DFMCRange}{}

\item{wind}{}

\item{windSD}{}

\item{windRange}{}

\item{fLine}{}

\item{moistureMultiplier}{}

\item{moistureSD}{}

\item{moistureRange}{}

\item{reps}{}

\item{Ms}{}

\item{Pm}{}

\item{Mr}{}

\item{vAir500}{}

\item{updateProgress}{}

\item{freeCores}{}

\item{l}{}
}
\value{
list
}
\description{
Models probabilistic fire behaviour across multiple ages on parallel cores
}
