% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{firePlant}
\alias{firePlant}
\title{firePlant}
\usage{
firePlant(
  dat,
  db.path = "out.plant.db",
  reps = 5,
  slope = 0,
  slopeSD = 2,
  slopeRange = 5,
  temp = 30,
  tempSD = 5,
  tempRange = 3,
  DFMC = 0.1,
  DFMCSD = 0.01,
  DFMCRange = 2,
  wind = 5,
  windSD = 1,
  windRange = 2,
  vAir500 = 2,
  moistureMultiplier = 1,
  moistureSD = 0.01,
  moistureRange = 1.5,
  fLine = 1000,
  leafVar = 0.1,
  updateProgress = NULL
)
}
\arguments{
\item{dat}{The results of frameDynTab}

\item{db.path}{Pathway to the database storing model results}

\item{reps}{Number of runs per age}

\item{slope}{Slope of the site (degrees)}

\item{slopeSD}{Standard deviation of slope}

\item{slopeRange}{Range of slope}

\item{temp}{Temperature of the site (degrees C)}

\item{tempSD}{Standard deviation of temperature}

\item{tempRange}{Range of temperature}

\item{DFMC}{Dead fuel moisture content (Proportion ODW)}

\item{DFMCSD}{Standard deviation of DFMC}

\item{DFMCRange}{Range of DFMC}

\item{wind}{Wind speed (kph)}

\item{windSD}{Standard deviation of wind speed}

\item{windRange}{Range of wind speed}

\item{vAir500}{Wind speed at 500m (kph)}

\item{moistureMultiplier}{Multiplies leaf moisture}

\item{moistureSD}{Standard deviation of leaf moisture}

\item{moistureRange}{Range of leaf moisture}

\item{fLine}{Length of the active fire front (m)}

\item{leafVar}{Variability in leaf thickness}

\item{updateProgress}{Function to update progress}
}
\description{
Models fire behaviour across a range of ages
from plant modelling
}
