% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{floraTarget}
\alias{floraTarget}
\title{floraTarget}
\usage{
floraTarget(Surf, Plant, Param = Param, targSp = "a", Test = 70)
}
\arguments{
\item{Surf}{The dataframe produced by the function 'summary',}

\item{Plant}{The dataframe produced by the function 'repFlame'.}

\item{Param}{A parameter dataframe used for FRaME,
such as produced using readLegacyParamFile}

\item{targSp}{Name of species to be watched for scorch}

\item{Test}{The temperature of the flora, default 70 degC}
}
\value{
dataframe
}
\description{
Scorch height using an isotherm, including target species

Calculates the height to which vegetation will be consumed,
and the height to a designated temperature isotherm reached for one second.

Output fields are:
Height - overall scorch height (m)
ns, e, m, c - height of consumption in each stratum (m)
b1, b2, b3, b4 - percentage of strata 1 (ns) to strata 4 (c) consumed
sc1, sc2, sc3, sc4 - percentage of strata 1 (ns) to strata 4 (c) scorched
d1, d2, d3, d4 - death (1) or survival (0) of standing foliage per stratum (50% or more scorch)
}
