% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestflames.R
\docType{package}
\name{forestflames}
\alias{forestflames}
\alias{forestflames-package}
\title{forestflames Forest flammability model.}
\description{
Simulation of the effect of vegetation structure and plant traits
on fire characteristics.
}
\references{
Zylstra et al. Title. PLOS (2016)
}

