% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{frameRisk}
\alias{frameRisk}
\title{frameRisk}
\usage{
frameRisk(
  Flora,
  Structure,
  default.species.params,
  a = a,
  weather,
  lightning = 0.05,
  fires = 5,
  smoulder = 24,
  Extinction = 0.15,
  hourStep = 3,
  slopeM = 6.7,
  slopeSD = 4.9,
  slopeLength = 391,
  rangeDir = 270,
  mRiver = 0.38,
  lRiver = 0.06,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  Test = 80,
  sensitive = 1,
  girdleH = c(),
  woodDensity = 1000,
  phloem = 0.01,
  moisture = 0.6,
  vAir500 = 2,
  barkDensity = 300,
  bark = 0.04,
  comBark = 700,
  resBark = 0,
  bMoisture = 0.5,
  distance = 0.1,
  trail = 10,
  var = 10,
  Altitude = 200,
  necT = 60,
  surfDecl = 10,
  minR = 1,
  targSp = "a",
  testN = 5
)
}
\arguments{
\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - A name for the site
species - the name of the species, which will call trait data from 
'default.species.params'
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - Percent composition of that species in the stratum. 
   If absent, all species will be considered equally.
base - base height of plant crowns (m)
he - he height of plant crowns (m)
ht - ht height of plant crowns (m)
top - top height of plant crowns (m)
w - width of plant crowns (m)
Hs - standard deviation of the top height of plant crowns (m)
Hr - range of the top height of plant crowns (m)
weight - weight in t/ha of fine dead organic material forming 
   the surface and suspNS layers
diameter - mean diameter of surface and suspNS litter in m}

\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. 
   Acceptable values are TRUE, FALSE, or blank, where the outcome 
   will be decided by the relative stratum heights.
nsR, eR, mR, cR. Species richness (number of species) in each stratum}

\item{default.species.params}{Plant traits database}

\item{a}{A unique identifier for the record being run}

\item{weather}{Formatted weather dataset}

\item{lightning}{Likelihood of lightning strike}

\item{fires}{Number of fires to model for each age}

\item{smoulder}{Time for which an ignition may smoulder (hours)}

\item{Extinction}{Extinction litter moisture content (Percent ODW)}

\item{hourStep}{Number of hours for which to model}

\item{slopeM}{Slope mean}

\item{slopeSD}{Slope standard deviation}

\item{slopeLength}{Distance from ridgeline to river}

\item{rangeDir}{Wind direction in relation to slope exposure: Up sun slope = 1, cross slope = 0, down sun slope = -1}

\item{mRiver}{Likelihood of encountering a medium river}

\item{lRiver}{Likelihood of encountering a large river}

\item{l}{Variability in leaf thickness}

\item{Ms}{Standard deviation of leaf moisture content}

\item{Pm}{Leaf moisture multiplier}

\item{Mr}{Leaf moisture range}

\item{Test}{Temperature at which to test for scorching (C)}

\item{sensitive}{Sensitivity of plants to scorching}

\item{girdleH}{Height at which to test for girdling}

\item{woodDensity}{Density of wood (kg/m^3)}

\item{phloem}{Thickness of phloem (m)}

\item{moisture}{Moisture content of wood}

\item{vAir500}{Multiplier for wind speed at 500hPa}

\item{barkDensity}{Density of bark (kg/m^3)}

\item{bark}{Bark thickness (m)}

\item{comBark}{Temperature of burning bark (C)}

\item{resBark}{Flame residence time on bark (s)}

\item{bMoisture}{Moisture content of bark}

\item{distance}{Distance between tree and flame start (m)}

\item{trail}{Modelling time after flame has passed (s)}

\item{var}{Variability in flame angle}

\item{Altitude}{Altitude of the site (m)}

\item{necT}{Temperature at which wood necrosis occurs (C)}

\item{surfDecl}{Decline in surface temperature with height (C/m)}

\item{minR}{Minimum ROS (m/h)}

\item{targSp}{Target species for scorching}

\item{testN}{Number of test runs}
}
\value{
list
}
\description{
Find likelihood and consequence of death for canopy trees
}
