% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{frameSummary}
\alias{frameSummary}
\title{Summary table of fire behaviour}
\usage{
frameSummary(flames, sites, ros, surface)
}
\arguments{
\item{flames}{The dataframe $FlameSummaries}

\item{sites}{The dataframe $Sites}

\item{ros}{The dataframe $ROS}

\item{surface}{The dataframe $SurfaceResults}
}
\value{
dataframe
}
\description{
Summarises FRaME generated fire behaviour by RepId
}
