% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{frameSurvey}
\alias{frameSurvey}
\title{frameSurvey}
\usage{
frameSurvey(
  dat,
  default.species.params,
  negEx = 1,
  max = 54.22,
  rate = 0.026,
  a = 3.35,
  b = 0.832,
  age = NA,
  surf = 10,
  density = 300,
  cover = 0.8,
  aQ = NA,
  bQ = NA,
  cQ = NA,
  maxNS = NA,
  rateNS = NA,
  wNS = 1,
  thin = TRUE,
  sLit = TRUE,
  dec = TRUE,
  sepSig = 0.001,
  messages = F,
  overlap = 0.5
)
}
\arguments{
\item{dat}{The dataframe containing the formatted field survey data}

\item{default.species.params}{Plant traits database}

\item{negEx}{Value determining the model used. 1 = olson, 2 = Burr}

\item{max}{Maximum weight of surface litter (t/ha)}

\item{rate}{Growth rate for surface litter in a negative exponential function}

\item{a}{Parameter in the Burr equation}

\item{b}{Parameter in the Burr equation}

\item{age}{Optional field with site age}

\item{surf}{Weight of surface litter in t/ha}

\item{density}{Wood density (kg.m-3)}

\item{cover}{Percent cover of suspended layer}

\item{aQ}{Parameter for a quadratic trend; leave as NA if trend is negative exponential}

\item{bQ}{Parameter for a quadratic trend; leave as NA if trend is negative exponential}

\item{cQ}{Parameter for a quadratic trend; leave as NA if trend is negative exponential}

\item{maxNS}{Parameter for a negative exponential trend; leave as NA if trend is quadratic}

\item{rateNS}{Parameter for a negative exponential trend; leave as NA if trend is quadratic}

\item{wNS}{Width of near surface patches (m)}

\item{thin}{Logical - TRUE uses plant self-thinning models in Dynamics,
otherwise plant cover remains at the highest point it has reached by that age}

\item{sLit}{Logical - TRUE allows surface litter to decline if the model does so, otherwise
the maximum value to that age is maintained}

\item{dec}{Logical - TRUE allows near surface surface litter to decline if the model does so, otherwise}

\item{sepSig}{Threshold for determining significance}

\item{messages}{T or F to display messages from component functions}

\item{overlap}{Either 'automatic', or threshold occurrence at which overlap is set to TRUE.
the maximum value to that age is maintained}
}
\value{
list
}
\description{
Processes field survey data into tables formatted for fire modelling
}
