% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{frameTables}
\alias{frameTables}
\title{frameTables}
\usage{
frameTables(
  dat,
  tr,
  age,
  propSamp = 0.75,
  transects = 10,
  sepSig = 0.1,
  rec = 1,
  propDead = 0,
  leafForm = "Flat",
  lwRat = 3,
  leafA = 0.002547,
  ram = 5,
  ignitionTemp = 260,
  moist = 1,
  G.C_rat = 3,
  C.C_rat = 0.1,
  deltaL = 0.46,
  hw = 0,
  sLitter = 15,
  diameter = 0.005,
  minCov = 1e-04
)
}
\arguments{
\item{dat}{The results of run_scm_collect}

\item{tr}{Table of input traits}

\item{age}{Years since disturbance}

\item{propSamp}{Values closer to 0 have more accurate ratios of components but miss some cohorts}

\item{transects}{More transects ensure more cohorts}

\item{sepSig}{Significance test for separation of strata}

\item{rec}{Number of the record}

\item{propDead}{Proportion of foliage dead}

\item{leafForm}{Flat or Round}

\item{lwRat}{Ratio of leaf length to width}

\item{leafA}{Area of a leaf in m2}

\item{ram}{Stem ramification}

\item{ignitionTemp}{Temperature of the endotherm (degC)}

\item{moist}{Leaf moisture (ratio moisture weight to dry weight)}

\item{G.C_rat}{Ratio of gaps to clumps of leaves}

\item{C.C_rat}{Ratio of clump to canopy size}

\item{deltaL}{Leaf density (g/cm3)}

\item{hw}{Environmental difference in plant height: width ratio}

\item{sLitter}{Weight of surface litter (t/ha)}

\item{diameter}{Mean diameter of surface litter pieces (m)}

\item{minCov}{Minimum cover for a species to be included}
}
\description{
Constructs parameter files for FRaME using
outputs from plant
}
