% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{growPlants}
\alias{growPlants}
\title{Grows a table of species to a given age, listing cover and variability}
\usage{
growPlants(Dynamics, Age = 10)
}
\arguments{
\item{Dynamics}{Table of growth models as output by floraDynamics}

\item{Age}{Age of the site in years}
}
\value{
dataframe
}
\description{
Grows a table of species to a given age, listing cover and variability
}
