% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{heDyn}
\alias{heDyn}
\title{heDyn}
\usage{
heDyn(
  dat,
  thres = 5,
  p = 0.05,
  species_col = "Species",
  grouping = "Age",
  pN = "Point",
  pntGap = 5,
  top = "top",
  he = "he"
)
}
\arguments{
\item{dat}{The dataframe containing the input data,}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{p}{The maximum allowable p value for a model}

\item{species_col}{The name of the column giving the name of the species (string)}

\item{grouping}{The name of the column that groups sites, eg by age or location (string)}

\item{pN}{The name of the column giving the number of the vertical transect (string)}

\item{pntGap}{The distance (m) between each vertical transect point}

\item{top}{Name of the top height field}

\item{he}{Name of the field with the height of the lower edge (he)}
}
\value{
dataframe
}
\description{
Builds models for top-he height allometry dynamics of surveyed Species
Species that are less common than the set threshold are combined as "Minor Species"
}
