% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threat.R
\name{hFauna}
\alias{hFauna}
\title{hFauna}
\usage{
hFauna(Shape = "Cylinder", Re)
}
\arguments{
\item{Shape}{Shape of the object, either 'Flat', 'Sphere' or 'Cylinder'}

\item{Re}{Reynolds number}
}
\value{
value
}
\description{
hFauna
}
