% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{hollow}
\alias{hollow}
\title{Fire risk for an animal sheltering in a wooden hollow}
\usage{
hollow(
  Surf,
  Plant,
  Height = 1,
  woodDensity = 700,
  barkDensity = 500,
  wood = 0.1,
  bark = 0.02,
  comBark = 700,
  resBark = 45,
  RH = 0.5,
  moisture = 0.2,
  bMoisture = 0.5,
  distance = 5,
  trail = 360,
  var = 10,
  Pressure = 1013.25,
  Altitude = 0,
  Dimension = 0.3,
  Area = 0.03,
  diameter = 6,
  surfDecl = 2,
  startTemp = 21,
  Shape = "Cylinder",
  updateProgress = NULL
)
}
\arguments{
\item{Surf}{The dataframe 'runs' exported from Monte Carlos as 'Summary.csv'}

\item{Plant}{The dataframe 'IP' exported from Monte Carlos as 'IP.csv'.}

\item{Height}{The height directly over ground (m) at which the species is expected to shelter from a fire.}

\item{woodDensity}{The density of wood in the tree or log housing the hollow (kg/m3)}

\item{barkDensity}{The density of bark in the tree or log housing the hollow (kg/m3)}

\item{wood}{The thickness of wood on the thinnest side of the hollow (m)}

\item{bark}{The thickness of bark on the thinnest side of the hollow (m)}

\item{comBark}{Temperature directly under the burning bark (C)}

\item{resBark}{Flame residence in the plant bark (s)}

\item{RH}{The relative humidity (0-1)}

\item{moisture}{The proportion oven-dry weight of moisture in the wood}

\item{bMoisture}{The proportion oven-dry weight of moisture in the bark}

\item{distance}{The furthest horizontal distance between the flame origin and the point (m)}

\item{trail}{The number of seconds to continue modelling after all flames have extinguished}

\item{var}{The angle in degrees that the plume spreads above/below a central vector;defaults to 10}

\item{Pressure}{Sea level atmospheric pressure (hPa)}

\item{Altitude}{Height above sea level (m)}

\item{Dimension}{The "Characteristic length" of the hollow (m)}

\item{Area}{The surface area of the thinnest side of the hollow (m^2)}

\item{diameter}{depth of the litter layer (mm)}

\item{surfDecl}{The slope of the surface (degrees)}

\item{startTemp}{The initial temperature of the hollow (C)}

\item{Shape}{The approximate shape of the hollow exterior - either "Flat", "Sphere", or "Cylinder"}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\value{
dataframe
}
\description{
Calculates the likelihood of mortality to an animal
caused by an approaching fire front
}
\details{
Utilises the output tables from 'threat' and 'radiation', and adds to these
the Reynolds Number, heat transfer coefficients, Newton's convective energy transfer coefficient,
and the temperature of the object each second.

Reynolds Number utilises a standard formulation (e.g. Gordon, N. T., McMahon, T. A. & Finlayson, B. L.
Stream hydrology: an introduction for ecologists. (Wiley, 1992))

Convective heat transfer coefficients use the widely adopted formulations of
Williams, F. A. Urban and wildland fire phenomenology. Prog. Energy Combust. Sci. 8, 317–354 (1982),
and Drysdale, D. An introduction to fire dynamics. (John Wiley and Sons, 1985)
utilising a Prandtl number of 0.7.

Finds animal mortality within a hollow based on the maximum tolerable temperature for a given
vapour pressure deficit, based on data from Lawrence, G. E.
Ecology of vertebrate animals in relation to chaparral fire in the Sierra Nevada foothills.
Ecology 47, 278–291 (1966)

Heat is transferred into the hollow using Fourier's Law

Thermal conductivity of bark is modelled as per Martin, R. E.
Thermal properties of bark. For. Prod. J. 13, 419–426 (1963)

Specific heat of bark is modelled using Kain, G., Barbu, M. C., Hinterreiter, S., Richter, K. & Petutschnigg, A.
Using bark as a heat insulation material. BioResources 8, 3718–3731 (2013)

Thermal conductivity of wood is modelled using an approach from Kollmann, F. F. P. & Cote, W. A.
Principles of wood science and technology I. Solid wood. (Springer-Verlag, 1968)

Evaporates water at 100 degrees C

Specific heat of wood is derived from an established empirical relationship in Volbehr, B.
Swelling of wood fiber. PhD Thesis. (University of Kiel, 1896)
}
