% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMC.R
\name{insol}
\alias{insol}
\title{Finds solar radiation reaching the ground surface}
\usage{
insol(
  lat = -32,
  hr = 12,
  month = "December",
  aspect = 0,
  slope = 0,
  LAI = 3,
  cloud = 0
)
}
\arguments{
\item{lat}{Latitude in degrees}

\item{hr}{Hour of the day, 0-24, decimal format}

\item{month}{Name of the month, capitalised (e.g. "January")}

\item{aspect}{Aspect in degrees}

\item{slope}{Slope in degrees}

\item{LAI}{Leaf Area Index}

\item{cloud}{Proportion cloud cover (0-1)}
}
\description{
Accounts for latitude, terrain, cloud cover, and leaf area index
}
