% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_functions.R
\name{isSpeciesKnown}
\alias{isSpeciesKnown}
\title{Checks if species names are defined in the default species parameters table.}
\usage{
isSpeciesKnown(names)
}
\arguments{
\item{names}{The names, whole or partial, to search for.}
}
\value{
A named boolean vector where names are search terms.
}
\description{
For each name (whole or partial) checks whether the name matches one, and only
one, record in the default species parameters table.
}

