% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earth.R
\name{kSoil}
\alias{kSoil}
\title{kSoil}
\usage{
kSoil(texture = "loam", saturation = 0.3, grain = "fine", unfrozen = 1)
}
\arguments{
\item{texture}{Soil texture}

\item{saturation}{Volumetric water content}

\item{grain}{Grain size}

\item{unfrozen}{Proportion of soil unfrozen}
}
\description{
Thermal conductivity of soil

Model drawn from Johansen, O.
Thermal conductivity of soils. PhD Thesis (University of Trondheim, 1971)
Modified by Farouki, O. Thermal properties of soils. (Trans Tech, 1986)

Porosity taken from Rawls, W. J., Brakensiek, D. L. & Saxton, K. E.
Estimation of soil water properties. Transactions of the ASAE 25, 1316–1320 & 1328 (1982).
}
