% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{LAIcommX}
\alias{LAIcommX}
\title{LAIcommX}
\usage{
LAIcommX(base.params, yu = 100, yl = 0)
}
\arguments{
\item{base.params}{Parameter input file}

\item{yu}{Upper height to measure (m)}

\item{yl}{Lower height to measure (m)}
}
\description{
Calculates LAI for a horizontal slice of a plant community
}
