% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{LAIp}
\alias{LAIp}
\title{LAIp}
\usage{
LAIp(base.params, sp = 1, yu = 100, yl = 0)
}
\arguments{
\item{base.params}{Parameter input file}

\item{sp}{Number of the species}

\item{yu}{Height at the top of the slice (m)}

\item{yl}{Height at the base of the slice (m)}
}
\value{
list
}
\description{
Calculates LAI for a slice of a plant
}
