% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{mRSE}
\alias{mRSE}
\title{Finds the RSE for the mean of a vector}
\usage{
mRSE(dat)
}
\arguments{
\item{vec}{A vector of the values being predicted}
}
\value{
value
}
\description{
Finds the RSE for the mean of a vector
}
