% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{olson}
\alias{olson}
\title{Models the weight of surface litter from time since fire 
using an Olson negative exponential function}
\usage{
olson(max = 54.22, growth = 0.026, age = 10)
}
\arguments{
\item{max}{Maximum weight (t/ha)}

\item{age}{The number of years since last fire}

\item{rate}{Rate of growth for a negative exponential function}
}
\value{
dataframe
}
\description{
Models the weight of surface litter from time since fire 
using an Olson negative exponential function
}
