% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{paramBuilder}
\alias{paramBuilder}
\title{paramBuilder}
\usage{
paramBuilder(site, Structure, Flora, default.species.params, a)
}
\arguments{
\item{site}{A dataframe with the six fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
slope - slope in degrees
wind - velocity in km/h
temp - ambient temperature deg. C
dfmc - moisture content of fine dead fuels in whole numbers (eg 0.1 for 10 percent)
litter - weight in t/ha of fine dead organic material forming the O horizon
diameter - mean diameter of surface litter in m
fline - the fireline length in m}

\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. Acceptable values
are t, f, or blank, where the outcome will be decided by the relative stratum heights.}

\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100 percent ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - Percent composition of that species in the stratum. If absent, all species will be considered equally
hc, he, ht, hp & w - canopy dimensions for that species (m)
clump - mean ratio of clump diameter to crown diameter
openness - proportion of plant canopy occupied by gaps between clumps}

\item{default.species.params}{Leaf traits database}

\item{a}{The record number for which to build the table}
}
\description{
Constructs parameter files from formatted datasets
}
