% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{parArboreal}
\alias{parArboreal}
\title{Helper function to run fire and impact model for arboreal fauna}
\usage{
parArboreal(r)
}
\arguments{
\item{r}{A single row from the Conditions table}
}
\value{
List of dataframes:
\enumerate{
\item Runs - detailed results for each replicate;
\item IP - summary of ignition paths in each plant;
\item Shelter_list - summary of mortalities associated with sheltering animals
\item Exposed_list - summary of mortalities associated with exposed animals
}
}
\description{
Helper function to run fire and impact model for arboreal fauna
}
