% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{parBurnP}
\alias{parBurnP}
\title{Internal function for fireDynamics
Calculate fire dynamics using probFire_Frame}
\usage{
parBurnP(n)
}
\arguments{
\item{a}{A unique identifier for the record being run}
}
\value{
list
}
\description{
Internal function for fireDynamics
Calculate fire dynamics using probFire_Frame
}
