% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{parBurnPW}
\alias{parBurnPW}
\title{Internal function for fireDynamics
Calculate fire dynamics using probFire_Frame from a weather dataset}
\usage{
parBurnPW(n)
}
\arguments{
\item{a}{A unique identifier for the record being run}
}
\value{
list
}
\description{
Internal function for fireDynamics
Calculate fire dynamics using probFire_Frame from a weather dataset
}
