% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{parBurnW}
\alias{parBurnW}
\title{Internal function for fireDynamics
Calculate fire dynamics using weatherSet_Frame}
\usage{
parBurnW(n)
}
\arguments{
\item{a}{A unique identifier for the record being run}
}
\value{
list
}
\description{
Internal function for fireDynamics
Calculate fire dynamics using weatherSet_Frame
}
