% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{parRisk}
\alias{parRisk}
\title{Internal function for riskDynamics}
\usage{
parRisk(a)
}
\arguments{
\item{a}{A unique identifier for the record being run}
}
\value{
list
}
\description{
Internal function for riskDynamics
}
