% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{pCover}
\alias{pCover}
\title{pCover}
\usage{
pCover(mods, sp, Age = 10)
}
\arguments{
\item{mods}{A table of models fit to species, format as per modCollector}

\item{sp}{The name of a species in the table mods}

\item{Age}{The age of the plant (years since fire)}
}
\value{
value
}
\description{
Predicts proportion cover at a given age

Selects model from tabled models per species
}
