% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threat.R
\name{phi}
\alias{phi}
\title{phi}
\usage{
phi(Fl, Fa, S, D, H)
}
\arguments{
\item{Fl}{Flame length (m)}

\item{Fa}{Flame angle (degrees)}

\item{S}{Slope angle (degrees)}

\item{D}{Horizontal distance from the flame origin to the point (m)}

\item{H}{Vertical distance from the flame origin to the point (m)}
}
\description{
Configuration or view factor

Calculates the configuration factor for radiative heat transfer as used
in AS3959 for bushfire risk assessment of built structures
}
