% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{plantRisk}
\alias{plantRisk}
\title{plantRisk}
\usage{
plantRisk(
  dynDat,
  a,
  ignitions = 1,
  hourStep = 3,
  Area = 10,
  weather,
  vAir = 25,
  wStDev = 5,
  tAir = 30,
  dfmc = 0.05,
  slope = 0,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  Test = 80,
  fireN = 1
)
}
\arguments{
\item{dynDat}{List of three dataframes: Flora, Structure & default.species.params}

\item{a}{A unique identifier for the record being run}

\item{ignitions}{Number of ignitions to model}

\item{hourStep}{Number of hours for which to model}

\item{Area}{Area of the site (ha)}

\item{weather}{Formatted weather dataset}

\item{vAir}{Wind speed (km/h)}

\item{wStDev}{Wind speed standard deviation}

\item{tAir}{Air temperature (C)}

\item{dfmc}{Dead fuel moisture content}

\item{slope}{Slope (degrees)}

\item{l}{Leaf thickness variability}

\item{Ms}{Standard deviation of leaf moisture content}

\item{Pm}{Leaf moisture multiplier}

\item{Mr}{Leaf moisture range}

\item{Test}{Temperature at which to test for scorching (C)}

\item{fireN}{Number of fires to model for each age}
}
\value{
list
}
\description{
Calculates risk (likelihood & consequence) for plants in a given climate & terrain
}
