% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{plantVarFrame}
\alias{plantVarFrame}
\title{Randomly modifies plant traits within defined ranges for non-deterministic predictions}
\usage{
plantVarFrame(
  base.params,
  Strata,
  Species,
  Flora,
  a,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.001
)
}
\arguments{
\item{base.params}{Parameter input table}

\item{Strata}{Strata descriptor table output by the function 'strata'}

\item{Species}{Species descriptor table output by the function 'species'}

\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100% ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - % composition or count of that species in the stratum. If absent, all species will be considered equally
hc, he, ht, hp & w - canopy dimensions for that species (m)
clump - mean ratio of clump diameter to crown diameter
openness - proportion of plant canopy occupied by gaps between clumps}

\item{a}{A unique identifier for the record being run}

\item{l}{Variation around input leaf dimensions}

\item{Ms}{Standard deviation of LFMC}

\item{Pm}{Multiplier of mean LFMC}

\item{Mr}{Truncates LFMC variability by +/- Mr * LFMC}
}
\value{
dataframe
}
\description{
Differs from plantVarS by using FRaME-updated tables
}
