% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{plumeAngle}
\alias{plumeAngle}
\title{Finds the angle of the convective plume at a given height}
\usage{
plumeAngle(base.params, Plume_velocity, Height)
}
\arguments{
\item{base.params}{A parameter table used by the FRaME workflow}

\item{Plume_velocity}{Velocity of convective airflow (m/s)}

\item{Height}{Height above ground (m)}
}
\value{
value
}
\description{
Finds the angle of the convective plume at a given height
}
