% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{probFire}
\alias{probFire}
\title{Models probabilistic fire behaviour}
\usage{
probFire(
  base.params,
  db.path = "out_mc.db",
  jitters,
  slope,
  slopeSD,
  slopeRange,
  temp,
  tempSD,
  tempRange,
  DFMC,
  DFMCSD,
  DFMCRange,
  wind,
  windSD,
  windRange,
  moistureMultiplier,
  moistureSD,
  moistureRange,
  heightSD,
  heightRange,
  leafVar,
  updateProgress = TRUE
)
}
\arguments{
\item{base.params}{Input parameter file}

\item{db.path}{Name of the exported database}

\item{jitters}{Number of repetitions for each row in the weather table}

\item{slope}{Mean slope (deg)}

\item{slopeSD}{Standard deviation of slope}

\item{slopeRange}{Truncates variability by +/- mean * range}

\item{temp}{Mean ambient temperature (deg.C)}

\item{tempSD}{Standard deviation of temp}

\item{tempRange}{Truncates variability by +/- mean * range}

\item{DFMC}{Mean DFMC (Percent ODW)}

\item{DFMCSD}{Standard deviation of DFMC}

\item{DFMCRange}{Truncates variability by +/- mean * range}

\item{wind}{Mean wind velocity (km/h)}

\item{windSD}{Standard deviation of wind velocity}

\item{windRange}{Truncates variability by +/- mean * range}

\item{moistureMultiplier}{Multiplies all LFMC values by this number}

\item{moistureSD}{Standard deviation of moisture}

\item{moistureRange}{Truncates variability by +/- mean * range}

\item{heightSD}{Standard deviation of plant height}

\item{heightRange}{Truncates variability by +/- mean * range}

\item{leafVar}{Variation around input leaf dimensions, equivalent to l}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\value{
dataframe
}
\description{
Models probabilistic fire behaviour
}
\examples{

SPECIFY INPUTS
record <- 1
data(site)
data(structure)
data(flora)
data(traits)
base.params <- paramBuilder(site, structure, flora, traits, record)

MODEL PROBABILISTIC FIRE BEHAVIOUR
probFire(base.params, db.path = "out.db", jitters = 50,
         slope = 0, slopeSD = 2, slopeRange = 5, 
         temp = 15, tempSD = 5, tempRange = 3,
         DFMC = 0.1, DFMCSD = 0.01, DFMCRange = 2, 
         wind = 5, windSD = 1, windRange = 2,
         moistureMultiplier = 1, moistureSD = 0.01, moistureRange = 1.5,
         heightSD = 2, heightRange = 1.41, 
         leafVar = 0.1,
         updateProgress = NULL)

LOAD AND ORGANISE RESULTS
res<-ffm_db_load("out.db")
}
