% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{probFire_Frame}
\alias{probFire_Frame}
\title{Models probabilistic fire behaviour using species-specific variability from FRaME tables}
\usage{
probFire_Frame(
  base.params,
  Structure,
  Flora,
  a,
  db.path = "out_mc.db",
  slope,
  slopeSD,
  slopeRange,
  temp,
  tempSD,
  tempRange,
  DFMC,
  DFMCSD,
  DFMCRange,
  wind,
  windSD,
  windRange,
  jitters = 50,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.001,
  updateProgress = NULL
)
}
\arguments{
\item{base.params}{Input parameter file}

\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. Acceptable values
are t, f, or blank, where the outcome will be decided by the relative stratum heights.
nsR, eR, mR, cR - maximum species richness recorded for each stratum}

\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100% ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - % composition or count of that species in the stratum. If absent, all species will be considered equally
hc, he, ht, hp & w - canopy dimensions for that species (m)
clump - mean ratio of clump diameter to crown diameter
openness - proportion of plant canopy occupied by gaps between clumps}

\item{db.path}{Name of the exported database}

\item{slope}{Mean slope (deg)}

\item{slopeSD}{Standard deviation of slope}

\item{slopeRange}{Truncates variability by +/- mean * range}

\item{temp}{Mean ambient temperature (deg.C)}

\item{tempSD}{Standard deviation of temp}

\item{tempRange}{Truncates variability by +/- mean * range}

\item{DFMC}{Mean DFMC (Percent ODW)}

\item{DFMCSD}{Standard deviation of DFMC}

\item{DFMCRange}{Truncates variability by +/- mean * range}

\item{wind}{Mean wind velocity (km/h)}

\item{windSD}{Standard deviation of wind velocity}

\item{windRange}{Truncates variability by +/- mean * range}

\item{jitters}{Number of repetitions for each row in the weather table}

\item{l}{Variation around input leaf dimensions}

\item{Ms}{Standard deviation of LFMC}

\item{Pm}{Multiplier of mean LFMC}

\item{Mr}{Truncates LFMC variability by +/- Mr * LFMC}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\value{
dataframe
}
\description{
Models probabilistic fire behaviour using species-specific variability from FRaME tables
}
\examples{

ADD EXAMPLE
}
