% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{randomiseHollow}
\alias{randomiseHollow}
\title{Randomise tree hollows}
\usage{
randomiseHollow(
  branchTrunkRat = 0.5,
  wallThickness_min = 0.001,
  wallThickness_max = 0.13,
  lengthHollow_min = 0.3,
  lengthHollow_max = 5,
  diameterHollow_min = 0.08,
  diameterHollow_max = 0.35,
  hollow_tilt_deg = 45,
  HTs = 10,
  HTmin = -10,
  tempDepression = 3.9,
  TDs = 2.4
)
}
\arguments{
\item{branchTrunkRat}{Ratio of hollows located in branches vs trunks (0-1)}

\item{wallThickness_min}{Minimum hollow wall thickness (m)}

\item{wallThickness_max}{Maximum hollow wall thickness (m)}

\item{lengthHollow_min}{Minimum hollow length (m)}

\item{lengthHollow_max}{Maximum hollow length (m)}

\item{diameterHollow_min}{Minimum hollow diameter (m)}

\item{diameterHollow_max}{Maximum hollow diameter (m)}

\item{hollow_tilt_deg}{Mean hollow tilt from horizontal (degrees)}

\item{HTs}{Standard deviation of hollow tilt (degrees)}

\item{HTmin}{Minimum hollow tilt (degrees)}

\item{tempDepression}{Mean temperature depression inside hollow (degrees C)}

\item{TDs}{Standard deviation of temperature depression (degrees C)}
}
\value{
Dataframe
}
\description{
Uses values from Hofman et al () to randomise hollow parameters for Monte Carlo simulations
}
