% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{remReps}
\alias{remReps}
\title{remReps}
\usage{
remReps(traits, printReplicates = TRUE)
}
\arguments{
\item{traits}{the table default.species.params}

\item{printReplicates}{Set TRUE if names of species removed are printed}
}
\value{
dataframe
}
\description{
Removes replicates in a formatted trait table
}
