% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{rich}
\alias{rich}
\title{rich}
\usage{
rich(
  dat,
  species_col = "Species",
  grouping = "Age",
  pN = "Point",
  thres = 5,
  pntGap = 5
)
}
\arguments{
\item{dat}{The dataframe containing the input data,}

\item{species_col}{The name of the column giving the name of the species (string)}

\item{grouping}{The name of the column that groups sites, eg by age or location (string)}

\item{pN}{The name of the column giving the number of the vertical transect (string)}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed
Species with lower cover are combined as "Minor Species"}

\item{pntGap}{The distance (m) between each vertical transect point}
}
\value{
dataframe
}
\description{
Finds the distribution of species richness at a point#'
}
