% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{richS}
\alias{richS}
\title{richS}
\usage{
richS(dat, thres = 0, pnts = 10, sepSig = 0.001)
}
\arguments{
\item{dat}{A dataframe listing plant species with columns describing crown dimensions using standardised names:
veg, pN, spName, base, top, he, ht, wid, Site, sN}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{pnts}{The number of points measured in a transect}

\item{sepSig}{Threshold for determining significance}
}
\value{
dataframe
}
\description{
Finds the distribution of species richness per stratum

Species that are less common than the set threshold are combined as "Minor Species"
}
