% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{riskDynamics}
\alias{riskDynamics}
\title{riskDynamics}
\usage{
riskDynamics(
  fireDat,
  tr,
  clim,
  lightning = 0.05,
  fires = 5,
  lat = -42.48,
  Altitude = 200,
  slopeM = 6.7,
  slopeSD = 4.9,
  slopeLength = 391,
  rangeDir = 270,
  mRiver = 0.38,
  lRiver = 0.06,
  cardinal = TRUE,
  smoulder = 24,
  Extinction = 0.15,
  m = 0.15,
  rholitter = 550,
  hourStep = 3,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  Test = 80,
  sensitive = 1,
  girdleH = c(),
  phloem = 0.01,
  moisture = 0.6,
  barkDensity = 300,
  bark = 0.04,
  comBark = 700,
  resBark = 0,
  bMoisture = 0.5,
  vAir500 = 2,
  testN = 5,
  distance = 2,
  trail = 1000,
  var = 10,
  necT = 60,
  surfDecl = 10,
  minR = 0.001,
  freeCores = 1
)
}
\arguments{
\item{fireDat}{List of three dataframes: Flora, Structure & default.species.params}

\item{tr}{A dataframe with the fields:}

\item{clim}{Formatted AM & PM weather conditions over a time sequence}

\item{lightning}{Number of lightning strikes per km2}

\item{fires}{Number of fires to model per age}

\item{lat}{Latitude (degrees)}

\item{Altitude}{Altitude (m.a.s.l.)}

\item{slopeM}{Mean slope (degrees)}

\item{slopeSD}{Standard deviation of slope (degrees)}

\item{slopeLength}{Distance from ridge to gully (m)}

\item{rangeDir}{Wind direction in relation to slope exposure: Up sun slope = 1, cross slope = 0, down sun slope = -1}

\item{mRiver}{Likelihood of encountering a medium river}

\item{lRiver}{Likelihood of encountering a large river}

\item{cardinal}{Logical, whether to use cardinal directions}

\item{smoulder}{Time for which an ignition may smoulder (hours)}

\item{Extinction}{Extinction litter moisture content (Percent ODW)}

\item{m}{Mean leaf moisture content}

\item{rholitter}{Density of litter (kg/m^3)}

\item{hourStep}{Number of hours for which to model}

\item{l}{Variability in leaf thickness}

\item{Ms}{Standard deviation of leaf moisture content}

\item{Pm}{Leaf moisture multiplier}

\item{Mr}{Leaf moisture range}

\item{Test}{Temperature at which to test for scorching (C)}

\item{sensitive}{Sensitivity of plants to scorching}

\item{girdleH}{Height at which to test for girdling}

\item{phloem}{Thickness of phloem (m)}

\item{moisture}{Moisture content of wood}

\item{barkDensity}{Density of bark (kg/m^3)}

\item{bark}{Bark thickness (m)}

\item{comBark}{Temperature of burning bark (C)}

\item{resBark}{Flame residence time on bark (s)}

\item{bMoisture}{Moisture content of bark}

\item{vAir500}{Multiplier for wind speed at 500hPa}

\item{testN}{Number of test runs}

\item{distance}{Distance between tree and flame start (m)}

\item{trail}{Modelling time after flame has passed (s)}

\item{var}{Variability in flame angle}

\item{necT}{Temperature at which wood necrosis occurs (C)}

\item{surfDecl}{Decline in surface temperature with height (C/m)}

\item{minR}{Minimum ROS (m/h)}

\item{freeCores}{Number of cores to leave free}
}
\value{
list
}
\description{
Calculates forest risk parameters over a growth series
using parallel ports
}
