% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{siteAP}
\alias{siteAP}
\title{Calculates atmospheric pressure at a given altitude}
\usage{
siteAP(
  Altitude,
  Pressure = 101325,
  ambientTemp = 281.65,
  lapseRate = 0.0065,
  g = 9.80665,
  Rgas = 287.058
)
}
\arguments{
\item{Altitude}{Altitude (m.a.s.l.)}

\item{Pressure}{Atmospheric pressure at sea level (Pa)}

\item{ambientTemp}{Ambient temperature (K)}

\item{lapseRate}{K/m (standard tropospheric lapse rate)}

\item{g}{Acceleration due to gravity (m/s^2)}

\item{Rgas}{J/(kg·K) for dry air}
}
\value{
Value
}
\description{
Calculates atmospheric pressure at a given altitude
}
