% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{speciesBuilder}
\alias{speciesBuilder}
\title{speciesBuilder}
\usage{
speciesBuilder(Flora, site, a)
}
\arguments{
\item{Flora}{A dataframe with the fields:
#' record - a unique, consecutively numbered identifier per site
#' species - the name of the species, which will call trait data from 'default.species.params'
#' moisture - the moisture content of the species in whole numbers (eg 1 for 100 percent ODW)
#' stratum - numeric value from 1 to 4, counting from lowest stratum
#' comp - Percent composition of that species in the stratum. If absent, all species will be considered equally
#' hc, he, ht, hp & w - canopy dimensions for that species (m)
#' clump - mean ratio of clump diameter to crown diameter
#' openness - proportion of plant canopy occupied by gaps between clumps}

\item{site}{A dataframe with the six fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
slope - slope in degrees
wind - velocity in km/h
temp - ambient temperature deg. C
dfmc - moisture content of fine dead fuels in whole numbers (eg 0.1 for 10 percent)
litter - weight in t/ha of fine dead organic material forming the O horizon
diameter - mean diameter of surface litter in m
fline - the fireline length in m}

\item{a}{The record number for which to build the table}
}
\value{
dataframe
}
\description{
Builds the dataframe species.values
}
