% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{standardiseNames}
\alias{standardiseNames}
\title{standardiseNames}
\usage{
standardiseNames(dat, pN, spName, base, top, he, ht, wid, Site, sN)
}
\arguments{
\item{dat}{The input table}

\item{pN}{A number specifying the point location of a vertical transect}

\item{spName}{Name of the species}

\item{base}{Base height of the plant crown (m)}

\item{top}{Top height of the plant crown (m)}

\item{he}{Lower edge height of the plant crown (m)}

\item{ht}{Upper height of the plant crown (m)}

\item{wid}{Width of the plant crown (m)}

\item{Site}{A number identifying the transect}

\item{sN}{A name identifying the transect}
}
\value{
table
}
\description{
Adds standardised names to a table of field data
}
