% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{stratify}
\alias{stratify}
\title{Arranges survey data into strata using k-means clustering}
\usage{
stratify(veg, pN = "Point", spName = "Species", pBase = "base", pTop = "top")
}
\arguments{
\item{veg}{A dataframe listing plant species with columns describing crown dimensions}

\item{pN}{The number of the point in the transect}

\item{spName}{Name of the field with the species name}

\item{pBase}{Name of the field with the base height}

\item{pTop}{Name of the field with the top height}
}
\value{
Dataframe
}
\description{
Data are stratified into 2-4 strata, then the number of strata 
with greatest significance is chosen.
Strata are sorted by top height and appended to the original data
}
