% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{stratList}
\alias{stratList}
\title{Stratifies multiple sites, rearranging strata 
into age-consistent trends}
\usage{
stratList(
  datAll,
  thres = 0,
  pnts = 10,
  base = "base",
  top = "top",
  p = 0.05,
  bTest = 3,
  maxiter = 1000
)
}
\arguments{
\item{datAll}{The dataframe containing the input data}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{pnts}{The number of points measured in a transect}

\item{base}{Name of the field with the base height}

\item{top}{Name of the field with the top height}

\item{p}{The maximum allowable p value for a model}

\item{bTest}{Multiples of mean + mRSE for which Burr & quadratic models can predict 
beyond the observed mean + standard deviation}

\item{maxiter}{The maximum number of iterations for model fitting}
}
\value{
dataframe
}
\description{
Input table requires the following fields:
Point - numbered point in a transect
Site - Site number for each point set
A field with the base height of the plants
A field with the top height of the plants
}
