% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{stratSite}
\alias{stratSite}
\title{stratSite}
\usage{
stratSite(
  dat,
  thres = 0,
  sepSig = 0.001,
  pN = "Point",
  spName = "spName",
  base = "base",
  top = "top",
  he = "he",
  ht = "ht"
)
}
\arguments{
\item{dat}{A dataframe listing plant species with columns describing crown dimensions using standardised names:
veg, pN, spName, base, top, he, ht, wid, Site, sN}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{sepSig}{Threshold for determining significance}

\item{pN}{Name of the column designating the point identifier}

\item{spName}{Name of the column designating the species name}

\item{base}{Name of the column with the canopy base height}

\item{top}{Name of the column with the canopy top height}

\item{he}{Name of the column with the he height}

\item{ht}{Name of the column with the ht height}
}
\value{
dataframe
}
\description{
Divides site data into consecutively numbered strata with
base and top heights

Species that are less common than the set threshold are combined as "Minor Species"
}
