% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{stratum}
\alias{stratum}
\title{Summary table of stratum results}
\usage{
stratum(flames, sites, ros, surface)
}
\arguments{
\item{flames}{The dataframe $FlameSummaries}

\item{sites}{The dataframe $Sites}

\item{ros}{The dataframe $ROS}

\item{surface}{The dataframe $SurfaceResults}
}
\value{
dataframe
}
\description{
Summarises FRaME generated fire behaviour by stratum and RepId
}
