% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threat.R
\name{threat}
\alias{threat}
\title{threat}
\usage{
threat(
  Surf,
  repFlame,
  Horizontal = 10,
  Height = 10,
  var = 10,
  Pressure = 1013.25,
  Altitude = 0,
  residence = 30,
  surfDecl = 2
)
}
\arguments{
\item{Surf}{The dataframe produced by the function 'summary',}

\item{repFlame}{The dataframe produced by the function 'repFlame'.}

\item{Horizontal}{The horizontal distance in metres from the flame origin to the point}

\item{Height}{The vertical distance in metres from the flame origin to the point}

\item{var}{The angle in degrees that the plume spreads above/below a central vector}

\item{Pressure}{Sea level atmospheric pressure (hPa)}

\item{Altitude}{Height above sea level (m)}

\item{residence}{Surface flame residence time}

\item{surfDecl}{Exponent describing the rate of post-front flame decay in surface litter}
}
\value{
dataframe
}
\description{
Calculates the convective and radiative heat from a flame that is 
incident upon a designated point

Finds the temperature, velocity, dynamic viscosity, atmospheric pressure, and density of a plume at a point
Gives the mean flame temperature (K) and emissive power of each flame

tempAir: Air temperature is modelled from dynamic flame segments using
Weber R.O., Gill A.M., Lyons P.R.A., Moore P.H.R., Bradstock R.A., Mercer G.N. (1995)
Modelling wildland fire temperatures. CALMScience Supplement, 4, 23–26.

cpAir: Specific heat of air is found from an empirical function fit to data from
Hilsenrath, J. et al.
Circular of the Bureau of Standards no. 564: tables of thermal properties of gases comprising tables of
thermodynamic and transport properties of air, argon, carbon dioxide, carbon monoxide hydrogen, nitrogen,
oxygen, and steam. (Department of Commerce, 1955), and
Kyle, B. G. Chemical and process thermodynamics. (Englewood Cliffs / Prentice Hall, 1984)

Density: Air density (kg/m^3) is found using the Ideal Gas Law.

presAtm: Air pressure (hPa) is calculated using the Barometric Formula with standard Values.

viscosity: Dynamic viscosity (10^-6 Pa.s) is calculated using an empirical relationship fit to a subset of
air pressures within an order of magnitude of sea level (R2 = 0.999), using data from
Kadoya, K., N, M. & Nagashima, A. Viscosity and thermal conductivity of dry air in the gaseous phase.
J. Phys. Chem. Ref. Data 14, 947–970 (1985)

Plume_velocity: Plume velocity (m/s) is calculated allowing for crosswinds, using
Oka Y., Sugawa O., Imamura T. (2008)
Correlation of temperature rise and velocity along an inclined fire plume axis in crosswinds.
Fire Safety Journal, 43, 391–400.

The Froude number is set to 1.5 as per Ma T.G., Quintiere J.G. (2003)
Numerical simulation of axi-symmetric fire plumes: Accuracy and limitations.
Fire Safety Journal, 38, 467–492.

flameTemp: Average flame temperature is integrated from the modelled temperature over the length of the flame

epsilon: The emissivity of a flame is calculated from mean flame thickness using
Àgueda, A., Pastor, E., Pérez, Y. & Planas, E.
Experimental study of the emissivity of flames resulting from the combustion of forest fuels.
Int. J. Therm. Sci. 49, 543–554 (2010)

phi: The configuration factor is calculated using the function phi, from
Tan, Z., Midgley, S. & Douglas, G.
A computerised model for bushfire attack assessment and its applications in bushfire protection planning.
Int. Congr. Model. Simul. Adv. Appl. Manag. Decis. Making, MODSIM05 538–545 (2005)
}
