% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuart_survey}
\alias{tuart_survey}
\title{Sample Tuart data}
\format{
a data frame with 9 columns and a row for each point of measurement
\describe{
\item{Site}{A numeric record of the site}
\item{SiteName}{A string name for the site}
\item{Point}{The point number in the transect, often the metre mark}
\item{Species}{The name of the species, consistent with other tables}
\item{base}{Crown base height hc (m)}
\item{top}{Crown top height hp (m)}
\item{he}{Crown he height (m)}
\item{ht}{Crown ht height (m)}
\item{width}{Crown width w (m)}

}
}
\source{
{Raw survey data collected using a 3d transect at Warrungup Spring reserve, April 2019}
}
\usage{
tuart_survey
}
\description{
A data frame containing sample field survey data for tuart woodland
}
\keyword{datasets}
