% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuart_traits}
\alias{tuart_traits}
\title{Tuart traits}
\format{
a data frame with 12 columns and a row for each species
\describe{
\item{name}{The name of the species, consistent with other tables}
\item{propDead}{Value between 0 & 1 giving the proportion of crown foliage that is dead}
\item{leafForm}{Either Flat or Round}
\item{leafThickness}{Thickness of the leaf (m)}
\item{leafWidth}{Width of the leaf at the widest point (m)}
\item{leafLength}{Length of the leaf (m)}
\item{leafSeparation}{Mean distance between leaf petioles on a 1st-order stem (m)}
\item{stemOrder}{Stem ramification or order}
\item{ignitionTemp}{Tempereature of the endotherm or piloted ignition temp (C)}
\item{moisture}{Ratio of moisture weight to oven-dry weight}
\item{G.C_rat}{Ratio of gaps to clumps, or crown space to foliage clumps}
\item{C.C_rat}{Ratio of individual clump area to crown area}

}
}
\source{
{Measurements from plants surveyed at Warrungup Spring reserve, April 2019}
}
\usage{
tuart_traits
}
\description{
A data frame containing the plant trait data for tuart woodland
}
\keyword{datasets}
