% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuart_weather}
\alias{tuart_weather}
\title{Tuart weather}
\format{
a data frame with 5 columns and a row for each time step
\describe{
\item{tm}{Consecutive integer to reference the time period}
\item{T}{Ambient air temperature (deg. C)}
\item{W}{Wind velocity (km/h)}
\item{DFMC}{Dead fuel moisture content (proportion oven dry weight)}
\item{time}{Hour of the day)}

}
}
\source{
{BOM AWS data for 16th May 2018, Mandurah weather station 009977. DFMC calculated using Matthews simple model}
}
\usage{
tuart_weather
}
\description{
A data frame containing the weather data for modelling in tuart woodland
}
\keyword{datasets}
