% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{updateSpecies}
\alias{updateSpecies}
\title{updateSpecies}
\usage{
updateSpecies(comm, tr)
}
\arguments{
\item{comm}{The output from stratify_community}

\item{tr}{An optional table of input traits}
}
\description{
Updates species names from an optional table
}
