% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{updateTraits}
\alias{updateTraits}
\title{updateTraits}
\usage{
updateTraits(
  traits,
  traitsNew,
  deleteReplicates = TRUE,
  printReplicates = TRUE,
  fill = TRUE
)
}
\arguments{
\item{traits}{Existing trait table}

\item{traitsNew}{Table containing new trait data}

\item{deleteReplicates}{If set to TRUE, retains the first mention of a species, then removes all following mentions}

\item{printReplicates}{If set to TRUE, prints the names of species that are repeated in the new table}

\item{fill}{If set to TRUE, fills empty numeric values with mean values of the genus}
}
\value{
table
}
\description{
Checks traits table for column names and adds data from a new table
}
