% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{buildFlora}
\alias{buildFlora}
\title{Constructs the table F_flora from formatted survey data}
\usage{
buildFlora(veg, surf = 20, sepSig = 0.001)
}
\arguments{
\item{veg}{A dataframe listing plant species with columns describing crown dimensions using standardised names:
veg, pN, spName, base, top, he, ht, wid, Site, sN}

\item{surf}{Weight of surface litter in t/ha}

\item{sepSig}{}
}
\value{
dataframe
}
\description{
Constructs the table F_flora from formatted survey data
}
