% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{buildFlora}
\alias{buildFlora}
\title{Constructs the table F_flora from formatted survey data}
\usage{
buildFlora(
  veg,
  pN = "Point",
  spName = "Species",
  pBase = "base",
  pTop = "top",
  hE = "he",
  hT = "ht",
  wid = "width",
  rec = "Site",
  sN = "SiteName",
  surf = 20,
  suspNS = 20,
  stratSig = 0.05
)
}
\arguments{
\item{veg}{The dataframe containing the input data}

\item{pN}{The number of the point in the transect}

\item{spName}{Name of the field with the species name}

\item{pBase}{Name of the field with the base height}

\item{pTop}{Name of the field with the top height}

\item{hE}{Name of the field with dimension he}

\item{hT}{Name of the field with dimension ht}

\item{wid}{Name of the field with dimension width}

\item{rec}{Name of the field with the record number}

\item{sN}{Optional field with a site name}

\item{surf}{Weight of surface litter in t/ha}

\item{suspNS}{Weight of suspended litter in t/ha}

\item{stratSig}{Value for test of significance}
}
\value{
dataframe
}
\description{
Constructs the table F_flora from formatted survey data
}
