% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{buildFloraP}
\alias{buildFloraP}
\title{Creates an 'F_flora' table for FRaME from plant modelling}
\usage{
buildFloraP(dat, age, rec = 1, moist = 1, litter = 15, diameter = 0.005)
}
\arguments{
\item{dat}{The output from stratify_community}

\item{age}{Years since disturbance}

\item{rec}{Number of the record}

\item{moist}{Leaf moisture (ratio moisture weight to dry weight)}

\item{litter}{Weight of surface litter (t/ha)}

\item{diameter}{Mean diameter of surface litter pieces (m)}
}
\description{
Creates an 'F_flora' table for FRaME from plant modelling
}
