% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{buildParams}
\alias{buildParams}
\title{buildParams}
\usage{
buildParams(
  Structure,
  Flora,
  default.species.params,
  a,
  fLine = 100,
  slope = 0,
  temp = 30,
  dfmc = 0.1,
  wind = 10
)
}
\arguments{
\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. 
   Acceptable values are TRUE, FALSE, or blank, where the outcome 
   will be decided by the relative stratum heights.
nsR, eR, mR, cR. Species richness (number of species) in each stratum}

\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - A name for the site
species - the name of the species, which will call trait data from 
'default.species.params'
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - Percent composition of that species in the stratum. 
   If absent, all species will be considered equally.
base - base height of plant crowns (m)
he - he height of plant crowns (m)
ht - ht height of plant crowns (m)
top - top height of plant crowns (m)
w - width of plant crowns (m)
Hs - standard deviation of the top height of plant crowns (m)
Hr - range of the top height of plant crowns (m)
weight - weight in t/ha of fine dead organic material forming 
   the surface and suspNS layers
diameter - mean diameter of surface and suspNS litter in m}

\item{default.species.params}{Plant traits database}

\item{a}{The number of the record for the site}

\item{fLine}{The fireline length in m}

\item{slope}{slope in degrees}

\item{temp}{Ambient temperature deg. C}

\item{dfmc}{Moisture content of fine dead fuels in whole numbers (eg 0.1 for 10 percent)}

\item{wind}{Velocity in km/h}
}
\description{
Constructs parameter files from imported tables
}
