% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{buildStructure}
\alias{buildStructure}
\title{Constructs the table F_structure from formatted survey data}
\usage{
buildStructure(veg, overlap = 0.5, sepSig = 0.001)
}
\arguments{
\item{veg}{A dataframe listing plant species with columns describing crown dimensions using standardised names:
veg, pN, spName, base, top, he, ht, wid, Site, sN}

\item{overlap}{Either 'automatic', or threshold occurrence at which overlap is set to TRUE.}

\item{sepSig}{Significance at which to recognise separate strata}
}
\value{
dataframe
}
\description{
Constructs the table F_structure from formatted survey data
}
