% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{buildStructure}
\alias{buildStructure}
\title{Constructs the table F_structure from formatted survey data}
\usage{
buildStructure(
  veg,
  pN = "Point",
  spName = "Species",
  pBase = "base",
  pTop = "top",
  rec = "Site",
  sN = "SiteName"
)
}
\arguments{
\item{veg}{The dataframe containing the input data}

\item{pN}{The number of the point in the transect}

\item{spName}{Name of the field with the species name}

\item{pBase}{Name of the field with the base height}

\item{pTop}{Name of the field with the top height}

\item{rec}{Name of the field with the record number}

\item{sN}{Optional field with a site name}
}
\value{
dataframe
}
\description{
Constructs the table F_structure from formatted survey data
}
