% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{buildStructureP}
\alias{buildStructureP}
\title{Creates an 'F_structure' table for FRaME from plant modelling}
\usage{
buildStructureP(dat, age, rec = 1)
}
\arguments{
\item{dat}{The output from stratify_community}

\item{age}{Years since disturbance}

\item{rec}{Number of the record}
}
\value{
dataframe
}
\description{
Creates an 'F_structure' table for FRaME from plant modelling
}
