% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{buildTraitsP}
\alias{buildTraitsP}
\title{Creates traits table for FRaME from plant modelling}
\usage{
buildTraitsP(
  dat,
  propDead = 0,
  leafForm = "Flat",
  lwRat = 3,
  leafA = 0.002547,
  ram = 5,
  ignitionTemp = 260,
  moist = 1,
  G.C_rat = 3,
  C.C_rat = 0.1,
  density = 0.46,
  lma
)
}
\arguments{
\item{dat}{The output from stratify_community}

\item{propDead}{Proportion of foliage dead}

\item{leafForm}{Flat or Round}

\item{leafA}{Area of a leaf in m2}

\item{ignitionTemp}{Temperature of the endotherm (degC)}

\item{moist}{Leaf moisture (ratio moisture weight to dry weight)}

\item{G.C_rat}{Ratio of gaps to clumps of leaves}

\item{C.C_rat}{Ratio of clump to canopy size}

\item{density}{Leaf density (g/cm3)}

\item{lma}{List of LMA values per species (kgm−2)}

\item{lwRatio}{Ratio of leaf length to width}
}
\description{
Leaf dimensions default to ausTraits mean leaf size
}
