% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{burnPrint}
\alias{burnPrint}
\title{Calculate the expected area burned}
\usage{
burnPrint(
  Flora,
  Structure,
  default.species.params,
  a,
  hourStep = 3,
  Area = 10,
  vAir = 25,
  wStDev = 5,
  tAir = 30,
  dfmc = 0.05,
  slope = 0,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  fireN = 1
)
}
\arguments{
\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - Name of the site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100% ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - % composition or count of that species in the stratum. If absent, all species will be considered equally
base, he, ht, top & w - canopy dimensions for that species (m)
Hs - Standard deviation of plant heights
Hr - Range of plant heights
weight - weight of litter (t/h)
diameter - mean diameter of litter (m)}

\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. Acceptable values
are t, f, or blank, where the outcome will be decided by the relative stratum heights.
nsR, eR, mR, cR - maximum species richness recorded for each stratum}

\item{default.species.params}{Leaf traits database}

\item{a}{A unique identifier for the record being run}

\item{hourStep}{Number of hours in each calculation timestep}

\item{Area}{Area of the study area (sq km)}

\item{l}{Variation around input leaf dimensions}

\item{Ms}{Standard deviation of LFMC}

\item{Pm}{Multiplier of mean LFMC}

\item{Mr}{Truncates LFMC variability by +/- Mr * LFMC}

\item{fireN}{Number of the fire being studied}
}
\value{
dataframe
}
\description{
Ember ignition likelihood calculated from 
Ellis P F 2011 Fuelbed ignition potential and bark morphology explain the 
notoriety of the eucalypt messmate “stringybark” for intense spotting Int. J. Wildl. Fire 20 897–907
}
