% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMC.R
\name{climDynamics}
\alias{climDynamics}
\title{Models input weather parameters from a climate dataset,
each age is modelled on a separate core}
\usage{
climDynamics(
  fireDat,
  clim,
  m = 0.15,
  slope = 5,
  slopeSD = 2,
  rholitter = 550,
  lat = -34.95,
  rangeDir = 270,
  cardinal = TRUE,
  freeCores = 1
)
}
\arguments{
\item{fireDat}{}

\item{clim}{}

\item{m}{}

\item{slope}{}

\item{slopeSD}{}

\item{rholitter}{}

\item{lat}{}

\item{rangeDir}{}

\item{cardinal}{}

\item{freeCores}{}

\item{litterW}{}
}
\value{
dataframe
}
\description{
Models input weather parameters from a climate dataset,
each age is modelled on a separate core
}
