% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMC.R
\name{climDynamics}
\alias{climDynamics}
\title{Models input weather parameters from a climate dataset,
each age is modelled on a separate core}
\usage{
climDynamics(
  fireDat,
  clim,
  m = 0.15,
  slope = 0,
  slopeSD = 0,
  rholitter = 550,
  lat = -34.95,
  lAngleAccounting = TRUE,
  rangeDir = 270,
  cardinal = TRUE,
  freeCores = 1
)
}
\arguments{
\item{fireDat}{}

\item{clim}{}

\item{m}{}

\item{slope}{(degrees)}

\item{slopeSD}{Standard deviation of the slope}

\item{rholitter}{}

\item{lat}{Latitude (degrees)}

\item{lAngleAccounting}{Set to TRUE to automatically modify leaf angle by number of strata}

\item{rangeDir}{Cardinal direction of the ridgelines - either Nth/Sth (0) or west/east (270)}

\item{cardinal}{}

\item{freeCores}{Number of cores to leave unused by the model}
}
\value{
dataframe
}
\description{
Models input weather parameters from a climate dataset,
each age is modelled on a separate core
}
