% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{collectTraitsP}
\alias{collectTraitsP}
\title{Creates traits table for FRaME from plant modelling}
\usage{
collectTraitsP(comm, tr, deltaL = 0.46)
}
\arguments{
\item{comm}{The output from stratify_community}

\item{tr}{A table of input traits}

\item{deltaL}{Leaf density (g/cm3)}
}
\description{
Reads traits from an input table
}
