% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{ctrlDiversity}
\alias{ctrlDiversity}
\title{Removes leaf trait diversity}
\usage{
ctrlDiversity(default.species.params)
}
\arguments{
\item{default.species.params}{Plant traits database}
}
\value{
dataframe
}
\description{
Removes leaf trait diversity
}
