% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{datClean}
\alias{datClean}
\title{Performs basic data checks on survey data, 
fixes or removes data and lists changes}
\usage{
datClean(veg, base = "base", top = "top", he = "he", ht = "ht")
}
\arguments{
\item{base}{Name of the field with the base height}

\item{top}{Name of the field with the top height}

\item{he}{Name of the field with dimension he}

\item{ht}{Name of the field with dimension ht}

\item{dat}{The dataframe to be cleaned}
}
\value{
dataframe
}
\description{
Performs basic data checks on survey data, 
fixes or removes data and lists changes
}
