% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DefaultSpeciesParams}
\alias{DefaultSpeciesParams}
\title{Species trait database.}
\format{A data frame where each row is a species; species name in
  the first column; trait values in subsequent columns.}
\usage{
DefaultSpeciesParams
}
\description{
Provides default parameter values for commonly referenced species. The 
following parameters are included (units, indicated where applicable,
accord with \code{\link{ParamInfo}}):
\itemize{
  \item propDead
  \item leafForm
  \item leafThickness (m)
  \item leafWidth (m)
  \item leafLength (m)
  \item leafSeparation (m)
  \item stemOrder
  \item ignitionTemp (degC)
}
Note that this does not include all mandatory parameters (e.g. crown dimensions).
}
\keyword{datasets}
