% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DefaultUnits}
\alias{DefaultUnits}
\title{Default units for model parameters.}
\format{A data frame with columns: section (values meta,
stratum, species); param (standard parameter label);
units (unit abbreviation).}
\usage{
DefaultUnits
}
\description{
Default units for model parameters.
}
\keyword{datasets}

