% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{driversFrame}
\alias{driversFrame}
\title{Models fire behaviour across ranged variables using species specific details}
\usage{
driversFrame(
  Flora,
  Structure,
  default.species.params,
  a,
  db.path = "out_mc.db",
  replicates = 3,
  winds = seq(0, 60, 5),
  slopes = 0,
  DFMCs = 0.05,
  tAir = 30,
  Pm = 1,
  Ms = 0.01,
  Mr = 0.01,
  l = 0.1,
  testN = 10,
  updateProgress = NULL,
  threshold = 1
)
}
\arguments{
\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100% ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - % composition or count of that species in the stratum. If absent, all species will be considered equally
hc, he, ht, hp & w - canopy dimensions for that species (m)
clump - mean ratio of clump diameter to crown diameter
openness - proportion of plant canopy occupied by gaps between clumps}

\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. Acceptable values
are t, f, or blank, where the outcome will be decided by the relative stratum heights.}

\item{default.species.params}{}

\item{a}{A unique identifier for the record being run}

\item{db.path}{Name of the exported database}

\item{replicates}{}

\item{winds}{}

\item{slopes}{List of slope values for testing}

\item{DFMCs}{List of DFMC values for testing}

\item{tAir}{}

\item{Pm}{}

\item{Ms}{}

\item{Mr}{}

\item{testN}{}

\item{updateProgress}{Progress bar for use in the dashboard}

\item{leafVar}{Variation around input leaf dimensions, equivalent to l}
}
\description{
Models fire behaviour across ranged variables using species specific details
}
