% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{dryside}
\alias{dryside}
\title{dryside}
\usage{
dryside(
  Surf,
  Plant,
  DBH = 1,
  Height = 0.1,
  woodDensity = 700,
  barkDensity = 500,
  bark = 0.04,
  comBark = 700,
  resBark = 45,
  phloem = 0.01,
  RH = 0.2,
  moisture = 0.6,
  bMoisture = 0.5,
  distance = 5,
  trail = 100,
  var = 10,
  diameter = 20,
  Pressure = 1013.25,
  Altitude = 0,
  startTemp = 25,
  necT = 60,
  surfDecl = 10,
  updateProgress = NULL
)
}
\arguments{
\item{Surf}{The dataframe produced by the function 'summary',}

\item{Plant}{The dataframe produced by the function 'repFlame'.}

\item{DBH}{Diameter at breast height (m)}

\item{Height}{Height of the tree (m)}

\item{woodDensity}{Density of the wood (kg/m^3)}

\item{barkDensity}{Density of the bark (kg/m^3)}

\item{bark}{Thickness of the bark (m)}

\item{comBark}{Conductivity of the bark (W/m/deg)}

\item{resBark}{Resistance of the bark (deg)}

\item{phloem}{Thickness of the phloem (m)}

\item{RH}{Relative humidity}

\item{moisture}{Moisture content of the wood}

\item{bMoisture}{Moisture content of the bark}

\item{distance}{Distance from the flame front (m)}

\item{trail}{Distance from the flame front to the tree (m)}

\item{var}{Variability of the flame front}

\item{diameter}{Diameter of the flame front (m)}

\item{Pressure}{Atmospheric pressure (hPa)}

\item{Altitude}{Altitude of the tree (m)}

\item{startTemp}{Starting temperature (deg C)}

\item{necT}{Temperature at which the tree is necrotic (deg C)}

\item{surfDecl}{Decline of the surface temperature (deg C)}

\item{updateProgress}{Function to update progress}
}
\value{
dataframe
}
\description{
Calculates likelihood of basal scarring on a tree

Assumes lee-side vortex at bole causes flame to lean backward to a distance 2.5*DBH
Estimate taken from Malcolm Gill A 1974 
Toward an understanding of fire scar formation: field observation and laboratory simulation 
For. Sci. 20 198–205
}
