% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{dryside}
\alias{dryside}
\title{Calculates likelihood of basal scarring on a tree}
\usage{
dryside(
  Surf,
  Plant,
  DBH = 1,
  Height = 0.1,
  woodDensity = 700,
  barkDensity = 500,
  bark = 0.04,
  comBark = 700,
  resBark = 45,
  phloem = 0.01,
  RH = 0.2,
  moisture = 0.6,
  bMoisture = 0.5,
  distance = 5,
  trail = 100,
  var = 10,
  diameter = 20,
  Pressure = 1013.25,
  Altitude = 0,
  startTemp = 25,
  necT = 60,
  surfDecl = 10,
  updateProgress = NULL
)
}
\arguments{
\item{Surf}{}

\item{Plant}{}

\item{DBH}{}

\item{Height}{}

\item{woodDensity}{}

\item{barkDensity}{}

\item{bark}{}

\item{comBark}{}

\item{resBark}{}

\item{phloem}{}

\item{RH}{}

\item{moisture}{}

\item{bMoisture}{}

\item{distance}{}

\item{trail}{}

\item{var}{}

\item{diameter}{}

\item{Pressure}{}

\item{Altitude}{}

\item{startTemp}{}

\item{necT}{}

\item{surfDecl}{}

\item{updateProgress}{}
}
\value{
dataframe
}
\description{
Assumes lee-side vortex at bole causes flame to lean backward to a distance 2.5*DBH
Estimate taken from Malcolm Gill A 1974 
Toward an understanding of fire scar formation: field observation and laboratory simulation 
For. Sci. 20 198–205
}
