% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMC.R
\name{Ema}
\alias{Ema}
\title{Ema}
\usage{
Ema(
  m,
  nelsonA = 5.2,
  nelsonB = -19,
  conLitter = 6e-04,
  tAir = 293.46,
  vAir = 3.52,
  pAir = 101180,
  insolation = 0,
  rhAir = 58
)
}
\arguments{
\item{m}{The starting moisture (proportion ODW)}

\item{nelsonA}{Constant from Nelson FMC model}

\item{nelsonB}{Constant from Nelson FMC model}

\item{conLitter}{Litter conductance (m/s). Using value from Matthews spreadsheet rather than paper}

\item{tAir}{Atmospheric temperature (K)}

\item{vAir}{Wind speed (m/s)}

\item{pAir}{Atmospheric pressure (Pa)}

\item{insolation}{Solar energy at the soil surface (kW/m2)}

\item{rhAir}{Relative humidity (Percent)}
}
\description{
Finds water vapour flux
Calculates RH at the litter surface from inversion of Nelson (1984) EMC equation
}
