% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f_flora}
\alias{f_flora}
\title{Flora data}
\format{
a data frame with 13 columns, and a row for each species per stratum, per site
\describe{
\item{record}{Consecutive integer to reference the row}
\item{site}{A string to name the row}
\item{species}{The name of the species, e.g. Eucalyptus rossii}
\item{moisture}{Moisture of live leaves (proportion oven dry weight) e.g 0.8}
\item{stratum}{Numbered from 1 (near surface) to 4 (canopy)}
\item{comp}{Integer giving the count of this species}
\item{base}{Height (m) of the base of the plant crown}
\item{he}{Height (m) of the lower edge of the plant crown}
\item{ht}{Height (m) of the upper edge of the plant crown}
\item{top}{Height (m) of the top of the plant crown}
\item{w}{Width (m) of the plant crown}
\item{openness}{Ratio of gaps between leaf clumps to clump diameter}
\item{clump}{Ratio of individual clump volume to crown volume}

}
}
\usage{
f_flora
}
\description{
A data frame containing the flora data for the model
}
\keyword{datasets}
