% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_functions.R
\name{ffm_assemble_table}
\alias{ffm_assemble_table}
\title{Assembles a parameter table from a list of elements.}
\usage{
ffm_assemble_table(lst)
}
\arguments{
\item{lst}{A list of table components.}
}
\value{
A standard five column parameter table.
}
\description{
This function takes a list of parameter table components as produced
by \code{\link{ffm_dissemble_table}} and combines them into a single
parameter table.
}
