% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_check_java}
\alias{ffm_check_java}
\title{Check for a locally installed Java runtime system}
\usage{
ffm_check_java(quiet = FALSE)
}
\value{
\code{TRUE} if a compatible version of Java is found; 
  \code{FALSE} otherwise.
}
\description{
Checks that Java version 1.8 or higher (aka Java 8) is installed 
on the local system and available from the command line. This is
required to run simulations with \code{\link{ffm_run}}.
}
