% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_db_functions.R
\name{ffm_db_load}
\alias{ffm_db_load}
\title{Database functions for the ffm package.}
\usage{
ffm_db_load(db.path)
}
\arguments{
\item{db.path}{The path and filename for the database.}
}
\value{
A named list of data frames corresponding to
  database tables.
}
\description{
Attempts to load the entire contents of a database into
the R workspace.

Given a path and filename for a database of model results,
this function connects to the database and loads the full
contents of all tables into a \code{list} of data frames.
An error is returned if the database cannot be opened.
}
