% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_functions.R
\name{ffm_dissemble_table}
\alias{ffm_dissemble_table}
\title{Deprecated function replaced by \code{ffm_disassemble_table}}
\usage{
ffm_dissemble_table(tbl)
}
\arguments{
\item{tbl}{The input parameter table in standard format.}
}
\value{
A named list of table components.
}
\description{
New code should use \code{\link{ffm_disassemble_table}}.
}
\seealso{
\code{\link{ffm_assemble_table}} for the reverse process.
}
