% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_functions.R
\name{ffm_param_table_summary}
\alias{ffm_param_table_summary}
\title{Returns a summary of the given parameter table.}
\usage{
ffm_param_table_summary(tbl)
}
\description{
Returns a summary of the given parameter table.
}

