% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_functions.R
\name{ffm_param_variance}
\alias{ffm_param_variance}
\title{Randomly perturbs values for selected numeric parameters in a table.}
\usage{
ffm_param_variance(tbl, max.prop = 0.1, labels = c("leafWidth",
  "leafLength", "leafThickness", "leafSeparation"), dist = c("normal",
  "beta"), beta.a = 1, beta.b = 1)
}
\description{
Given an input parameters table, this function adds random noise to 
specified numeric parameters and returns an updated copy of the table.
Noise can be based on a Normal (Gaussian) or Beta distribution. The 
maximum variation is controlled by the \code{max.prop} argument. 
When adding Gaussian noise, a random value is drawn from a Normal 
distribution with a mean of zero and standard deviation equal to
\code{max.prop / 3}. This results in 99% of percent of random values
being within plus or minus \code{max.prop} times the input value.
Any random perturbations outside this range will be discarded and 
re-drawn (so effectively working with a truncated Normal distribution).
}

