% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLegacyParamFile.R
\name{ffm_read_legacy_params}
\alias{ffm_read_legacy_params}
\title{Reads a legacy format parameters file}
\usage{
ffm_read_legacy_params(path, raw = FALSE)
}
\arguments{
\item{path}{The full or relative path to the parameters file.}

\item{raw}{(logical) Whether to return raw parsed data. This can be useful
for diagnosing problems with reading a particular parameter file. It should
be set to \code{FALSE} (default) to return a standard format table of
parameters.}
}
\value{
If \code{raw} is \code{TRUE}, a data frame of parsed data
  in raw form (useful for testing).
  If \code{raw} is \code{FALSE} (the default), a data frame in a form 
  suitable for use with \code{\link{ffm_run}}.
}
\description{
The legacy parameters file format was designed for use with a C++
implementation of the model. It does not conform to any more general format
(e.g. json or yaml) and, to be read correctly, relies on implicit assumptions
about the relationship between variables. For new parameter sets it is
preferable to compose them as tables in R or elsewhere and save them as
CSV-format files.
}
\examples{
\dontrun{
# Import a legacy format parameter file and save it as a CSV format file
# in standard tabular parameter format.
params <- ffm_read_legacy_params("old_params.txt")
ffm_write_params(params, file = "beautiful_params.csv")
}

}
