% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run}
\alias{ffm_run}
\title{Runs the model with the given site.}
\usage{
ffm_run(x, db, ...)
}
\arguments{
\item{x}{Site data (parameter set) as one of the following:
\itemize{
  \item a ScalaInterpreterReference object created with \code{\link{ffm_site}}
  \item a a matrix or data.frame valid for input to \code{ffm_site}
  \item a parameter generating function.
}}

\item{db}{Output database}

\item{additional}{arguments (see specific ffm_run functions for details)}
}
\description{
Runs the model with the given site.
}
\seealso{
\code{\link{ffm_run.ScalaInterpreterReference}}

\code{\link{ffm_run.data.frame}}

\code{\link{ffm_run.matrix}}

\code{\link{ffm_run.function}}
}

