% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run}
\alias{ffm_run}
\title{Run the model with the given site.}
\usage{
ffm_run(site)
}
\arguments{
\item{The}{site as a ScalaInterpreterReference object.}
}
\description{
Run the model with the given site.
}

