% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run}
\alias{ffm_run}
\title{Run the model with the given site.}
\usage{
ffm_run(x)
}
\arguments{
\item{x}{Site data as a ScalaInterpreterReference object
created with \code{\link{ffm_site}} or a matrix or data.frame
valid for input to \code{ffm_site}.}
}
\description{
Run the model with the given site.
}

