% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run}
\alias{ffm_run}
\title{Runs the model with the given site.}
\usage{
ffm_run(x, db, ...)
}
\arguments{
\item{x}{Site data (parameter set) as one of the following:
\itemize{
  \item a ScalaInterpreterReference object created with \code{\link{ffm_site}}
  \item a a matrix or data.frame valid for input to \code{ffm_site}
  \item a parameter generating function.
}}

\item{db}{Output database object or character string for path and filename.}

\item{additional}{arguments (see specific ffm_run functions for details)}
}
\description{
The destination database for model results (\code{db} argument) can be specified 
as either a database object, created with \code{\link{ffm_create_database}}, or
a character string for the path and filename.  If \code{db} is an object, results
are sent to the database which is then left open and available for use by subsequent
runs. If \code{db} is a character string for path and filename, a new database is 
created, written to, and then closed.
}
\seealso{
\code{\link{ffm_run.ScalaInterpreterReference}}

\code{\link{ffm_run.data.frame}}

\code{\link{ffm_run.matrix}}

\code{\link{ffm_run.function}}
}

