% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm_api_functions.R
\name{ffm_run_command}
\alias{ffm_run_command}
\title{Composes a command line call to run a simulation}
\usage{
ffm_run_command(param.path, db.path, db.recreate = FALSE,
  runtime.class = "ffm.runner.CSVRunner")
}
\arguments{
\item{param.path}{Path and name of a CSV-format file of simulation 
parameters.}

\item{db.path}{Path and name of the output SQLite database file.}

\item{db.recreate}{A logical value indicating, if the file specified 
by \code{db.path} exists, whether to recreating the database (\code{TRUE})
or append to it (\code{FALSE}). Default is to append.}

\item{runtime.class}{The name of the runtime class in the Scala model
application. Unless you have modified the Scala application this should
be left as the default value: \code{ffm.runner.CSVRunner}.}
}
\value{
The command line instruction as a one-element character vector.
}
\description{
This function is used by \code{\link{ffm_run}} when calling the Scala model
application via the base R \code{\link[base]{system}} function. The command
line string passed to \code{system} invokes the local Java runtime and
includes arguments for the necessary Java libraries (jar files included with
the package), the CSV-format file of input parameters, the path to the output
database, and (optionally) a flag indicating that the database should be
recreated if the file exists.
}
