% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_params_functions.R
\name{ffm_valid_param}
\alias{ffm_valid_param}
\title{Checks that a parameter label is valid.}
\usage{
ffm_valid_param(param, section = NULL, no.match.error)
}
\arguments{
\item{param}{The parameter label.}

\item{section}{If not \code{NULL}, one of site, stratum or species.}

\item{no.match.error}{If \code{TRUE} an error results when a unique
match is not found for the input label; if \code{FALSE} (default)
the function returns NULL when there is no unique match.}
}
\value{
\code{TRUE} is the label is valid.
}
\description{
A parameter label is valid if it is present in the \code{\link{ParamInfo}}
table. The comparison ignores case and white-space. If a non-NULL value is
supplied for \code{section} the check is restricted to parameter labels
in that section (site, stratum or species).
}

