% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{fireDynamics}
\alias{fireDynamics}
\title{Models fire behaviour across multiple ages on parallel cores
Uses either weatherSet_Frame or probFire_Frame}
\usage{
fireDynamics(
  fireDat,
  analysis = "Probabilistic",
  weather,
  slope = 5,
  slopeSD = 2,
  slopeRange = 5,
  temp = 30,
  tempSD = 5,
  tempRange = 3,
  DFMC = 0.1,
  DFMCSD = 0.01,
  DFMCRange = 2,
  wind = 10,
  windSD = 5,
  windRange = 5,
  fLine = 1000,
  moistureMultiplier = 1,
  moistureSD = 0.01,
  moistureRange = 1.5,
  reps = 5,
  leafVar = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  vAir500 = 2,
  testN = 5,
  updateProgress = NULL,
  threshold = 1,
  freeCores = 1
)
}
\arguments{
\item{fireDat}{A list containing the datasets Flora, Structure, and default.species.params}

\item{analysis}{Type of analysis, either "Weather" to use weatherSet_Frame, or "Probabilistic" to use probFire_Frame}

\item{weather}{A dataframe for use with weatherSet_Frame, with the five fields:
tm - Sequential numbering of the records
T - Air temperature (deg C)
W - Wind velocity (km/h)
DFMC - Dead fuel moisture content (proportion ODW)
record - A unique number for each age corresponding to the same fields in fireDat}

\item{slope}{Mean slope (degrees)}

\item{slopeSD}{Standard deviation of the slope (degrees)}

\item{slopeRange}{Range of slope (degrees)}

\item{temp}{}

\item{tempSD}{}

\item{tempRange}{}

\item{DFMC}{}

\item{DFMCSD}{}

\item{DFMCRange}{}

\item{wind}{}

\item{windSD}{}

\item{windRange}{}

\item{fLine}{Fireline length (m)}

\item{moistureMultiplier}{}

\item{moistureSD}{}

\item{moistureRange}{}

\item{reps}{Number of repetitions for each set of weather conditions}

\item{leafVar}{}

\item{Ms}{}

\item{Pm}{}

\item{Mr}{}

\item{vAir500}{Multiplier of wind speed to estimate 500 hPA wind}

\item{testN}{}

\item{updateProgress}{Progress bar for use in the dashboard}

\item{threshold}{Minimum allowable height for canopy (m)}

\item{freeCores}{Number of cores to leave free for other processes}
}
\value{
list
}
\description{
Models fire behaviour across multiple ages on parallel cores
Uses either weatherSet_Frame or probFire_Frame
}
