% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{firePlant}
\alias{firePlant}
\title{Models fire behaviour across a range of ages
from plant modelling}
\usage{
firePlant(
  dat,
  db.path = "out.plant.db",
  reps = 5,
  slope = 0,
  slopeSD = 2,
  slopeRange = 5,
  temp = 30,
  tempSD = 5,
  tempRange = 3,
  DFMC = 0.1,
  DFMCSD = 0.01,
  DFMCRange = 2,
  wind = 5,
  windSD = 1,
  windRange = 2,
  vAir500 = 2,
  moistureMultiplier = 1,
  moistureSD = 0.01,
  moistureRange = 1.5,
  fLine = 1000,
  leafVar = 0.1,
  updateProgress = NULL
)
}
\arguments{
\item{dat}{The results of frameDynTab}

\item{db.path}{}

\item{reps}{}

\item{slope}{}

\item{slopeSD}{}

\item{slopeRange}{}

\item{temp}{}

\item{tempSD}{}

\item{tempRange}{}

\item{DFMC}{}

\item{DFMCSD}{}

\item{DFMCRange}{}

\item{wind}{}

\item{windSD}{}

\item{windRange}{}

\item{moistureMultiplier}{}

\item{moistureSD}{}

\item{moistureRange}{}

\item{fLine}{Length of the active fire front (m)}

\item{leafVar}{}

\item{updateProgress}{}
}
\description{
Models fire behaviour across a range of ages
from plant modelling
}
