% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{firePlant}
\alias{firePlant}
\title{Models fire behaviour across a range of ages
from plant modelling}
\usage{
firePlant(
  dat,
  db.path = "out.plant.db",
  reps = 5,
  slope = 0,
  slopeSD = 2,
  slopeRange = 5,
  temp = 30,
  tempSD = 5,
  tempRange = 3,
  DFMC = 0.1,
  DFMCSD = 0.01,
  DFMCRange = 2,
  wind = 5,
  windSD = 1,
  windRange = 2,
  moistureMultiplier = 1,
  moistureSD = 0.01,
  moistureRange = 1.5,
  fLine = 100,
  leafVar = 0.1,
  updateProgress = NULL
)
}
\arguments{
\item{dat}{The results of frameDynTab}

\item{max}{Maximum years since disturbance}

\item{interval}{Time interval for sampling}

\item{rec}{Number of the record}

\item{sample}{Proportion of cohorts to test (0-1)}

\item{transects}{Number of repeats for each sample}

\item{openness}{Ratio of gaps to clumps of leaves}

\item{clump}{Ratio of clump frontal area diameter to crown frontal area diameter}

\item{litter}{Weight of surface litter (t/ha)}

\item{diameter}{Mean diameter of surface litter pieces (m)}

\item{propDead}{Proportion of foliage dead}

\item{leafForm}{Flat or Round}

\item{lwRatio}{Ratio of leaf length to width}

\item{leafA}{Area of a leaf in m2}

\item{ignitionTemp}{Temperature of the endotherm (degC)}

\item{moist}{Leaf moisture (ratio moisture weight to dry weight)}

\item{G.C_rat}{Ratio of gaps to clumps of leaves}

\item{C.C_rat}{Ratio of clump to canopy size}

\item{lat}{Latitude (degrees)}

\item{map}{Mean annual precipitation (mm)}

\item{mat}{Mean annual temperature (degC)}

\item{deltaL}{Leaf density (g/cm3)}

\item{lma}{List of LMA values per species (kgm−2)}
}
\description{
Models fire behaviour across a range of ages
from plant modelling
}
