% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{floraDynamics}
\alias{floraDynamics}
\title{floraDynamics}
\usage{
floraDynamics(
  dat,
  thres = 5,
  species_col = "Species",
  grouping = "Age",
  pN = "Point",
  base = "base",
  top = "top",
  he = "he",
  ht = "ht",
  width = "wid",
  pntGap = 5,
  p = 0.01,
  bTest = 2,
  cTest = 10,
  maxiter = 1000,
  Sr = 0,
  Sk = 0,
  Sa = 0,
  Sb = 0,
  Sc = 0,
  NSr = 0,
  NSk = 0,
  NSa = 0,
  NSb = 0,
  NSc = 0
)
}
\arguments{
\item{dat}{The dataframe containing the input data,}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{species_col}{The name of the column giving the name of the species (string)}

\item{grouping}{The name of the column that groups sites, eg by age or location (string)}

\item{pN}{The name of the column giving the number of the vertical transect (string)}

\item{base}{Name of the base height field}

\item{top}{Name of the field containing the height of the tree canopy}

\item{he}{Name of the field with dimension he}

\item{ht}{Name of the field with dimension ht}

\item{width}{Name of the field containing the width of the tree canopy}

\item{pntGap}{The distance (m) between each vertical transect point}

\item{p}{The maximum allowable p value for a model}

\item{bTest}{Multiples of mean + mRSE for which Burr & quadratic models can predict}

\item{cTest}{Multiples of mean + mRSE for which cover models can predict}

\item{maxiter}{The maximum number of iterations for model fitting}

\item{Sr}{Rate of increase for surface litter in a negative exponential curve}

\item{Sk}{Asymptote for surface litter in a negative exponential curve}

\item{Sa}{Parameter for quadratic modelling of surface litter weight}

\item{Sb}{Parameter for quadratic modelling of surface litter weight}

\item{Sc}{Parameter for quadratic modelling of surface litter weight}

\item{NSr}{Rate of increase for NS fuels in a negative exponential curve}

\item{NSk}{Asymptote for NS fuels in a negative exponential curve}

\item{NSa}{Parameter for quadratic modelling of NS fuels weight}

\item{NSb}{Parameter for quadratic modelling of NS fuels weight}

\item{NSc}{Parameter for quadratic modelling of NS fuels weight}
}
\value{
dataframe
}
\description{
Builds the plant growth table used in dynamic modelling

Models plant growth from field data in a series of models,
selecting the best model and providing error statistics.

Input table requires the following fields:
Point - numbered point in a transect
Species - name of the surveyed Species
Age - age of the site since the triggering disturbance

Species that are less common than the set threshold are combined as "Minor Species"
}
