% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\docType{_PACKAGE}
\name{frame}
\alias{frame}
\title{frame: Fire Research and Modelling Environment}
\description{
This package provides an interface to the Forest Flammability Model of
Zylstra et al, adding functions for complex fire simulation and first-order 
fire effects modelling. It provides  the primary functions required to prepare input 
parameters, run model simulations, and query the output database of model results,
as well as introducing complex simulations. First-order effects models are provided
for plant canopies, stems, and soil effects.
}
\details{
The package includes a distribution of the model and its dependencies in the
form of Java jar files. The only additional requirement is an installation
of Java version 1.8 (aka Java 8) or higher on the host system.
}
\references{
Zylstra, P. J. (2011) Forest flammability: modelling and managing a complex system. 
PhD Thesis, (University of NSW). doi:10.13140/2.1.3722.0166

Zylstra, P. J., Bradstock, R. A., Bedward, M., Penman, T. D., Doherty, M. D., 
Weber, R. O., … Cary, G. J. (2016). Biophysical mechanistic modelling quantifies 
the effects of plant traits on fire severity: species, not surface fuel loads 
determine flame dimensions in eucalypt forests. PLoS One, 11(8), e0160715. 

Zylstra P. J. (2021) Linking fire behaviour and its ecological effects to 
plant traits, using FRaME in R. Methods Ecol. Evol., 12(8), 1365-1378

Zylstra, P. J. (2023) Quantifying the direct fire threat to a critically endangered 
arboreal marsupial using biophysical, mechanistic modelling. Austral Ecol. 48, 266–288.
}
