% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\docType{package}
\name{frame}
\alias{frame}
\alias{frame-package}
\title{frame: Fire research and modelling environment}
\description{
This package provides an interface to the forest flammability model of
Zylstra et al. It provides functions to prepare input parameters, run model
simulations, and query the output database of model results. The package
includes a distribution of the model and its dependencies in the
form of Java jar files. The only additional requirement is an installation
of Java version 1.8 (aka Java 8) or higher on the host system.
}
\references{
Zylstra et al. PLOS (2016)
}
