% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{frameDynTab}
\alias{frameDynTab}
\title{Constructs an age sequence of parameter files 
built from plant modelling}
\usage{
frameDynTab(
  dat,
  tr,
  breaks = c(20, 50, 200),
  interval = c(2, 5, 10),
  propSamp = 1,
  transects = 1,
  propDead = 0,
  leafForm = "Flat",
  lwRat = 3,
  leafA = 0.002547,
  ram = 5,
  ignitionTemp = 260,
  moist = 1,
  G.C_rat = 3,
  C.C_rat = 0.1,
  deltaL = 0.46,
  hw = 0,
  mat = 17,
  diameter = 0.005
)
}
\arguments{
\item{dat}{The results of run_scm_collect}

\item{tr}{An optional table of input traits}

\item{breaks}{List of time periods for each sampling interval, length 3.}

\item{interval}{List of sampling intervals, length 3.}

\item{propSamp}{Proportion of cohorts to test (0-1)}

\item{transects}{Number of repeats for each sample}

\item{propDead}{Proportion of foliage dead}

\item{leafForm}{Flat or Round}

\item{lwRat}{}

\item{leafA}{Area of a leaf in m2}

\item{ram}{}

\item{ignitionTemp}{Temperature of the endotherm (degC)}

\item{moist}{Leaf moisture (ratio moisture weight to dry weight)}

\item{G.C_rat}{Ratio of gaps to clumps of leaves}

\item{C.C_rat}{Ratio of clump to canopy size}

\item{deltaL}{Leaf density (g/cm3)}

\item{hw}{}

\item{diameter}{Mean diameter of surface litter pieces (m)}
}
\description{
Constructs an age sequence of parameter files 
built from plant modelling
}
