% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{frameRisk}
\alias{frameRisk}
\title{Find likelihood and consequence of death for canopy trees}
\usage{
frameRisk(
  Flora,
  Structure,
  default.species.params,
  a = a,
  weather,
  lightning = 0.05,
  fires = 5,
  smoulder = 24,
  Extinction = 0.15,
  hourStep = 3,
  slopeM = 6.7,
  slopeSD = 4.9,
  slopeLength = 391,
  rangeDir = 270,
  mRiver = 0.38,
  lRiver = 0.06,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  Test = 80,
  sensitive = 1,
  girdleH = c(),
  woodDensity = 1000,
  phloem = 0.01,
  moisture = 0.6,
  vAir500 = 2,
  barkDensity = 300,
  bark = 0.04,
  comBark = 700,
  resBark = 0,
  bMoisture = 0.5,
  distance = 0.1,
  trail = 10,
  var = 10,
  Altitude = 200,
  necT = 60,
  surfDecl = 10,
  minR = 1,
  targSp = "a",
  testN = 5
)
}
\arguments{
\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - A name for the site
species - the name of the species, which will call trait data from 
'default.species.params'
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - Percent composition of that species in the stratum. 
   If absent, all species will be considered equally.
base - base height of plant crowns (m)
he - he height of plant crowns (m)
ht - ht height of plant crowns (m)
top - top height of plant crowns (m)
w - width of plant crowns (m)
Hs - standard deviation of the top height of plant crowns (m)
Hr - range of the top height of plant crowns (m)
weight - weight in t/ha of fine dead organic material forming 
   the surface and suspNS layers
diameter - mean diameter of surface and suspNS litter in m}

\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. 
   Acceptable values are TRUE, FALSE, or blank, where the outcome 
   will be decided by the relative stratum heights.
nsR, eR, mR, cR. Species richness (number of species) in each stratum}

\item{default.species.params}{Plant traits database}

\item{a}{A unique identifier for the record being run}

\item{weather}{}

\item{lightning}{}

\item{fires}{Number of fires to model for each age}

\item{smoulder}{}

\item{Extinction}{}

\item{hourStep}{}

\item{slopeM}{}

\item{slopeSD}{}

\item{slopeLength}{}

\item{rangeDir}{}

\item{mRiver}{}

\item{lRiver}{}

\item{l}{}

\item{Ms}{}

\item{Pm}{}

\item{Mr}{}

\item{Test}{}

\item{sensitive}{}

\item{girdleH}{}

\item{woodDensity}{}

\item{phloem}{}

\item{moisture}{}

\item{vAir500}{}

\item{barkDensity}{}

\item{bark}{}

\item{comBark}{}

\item{resBark}{}

\item{bMoisture}{}

\item{distance}{}

\item{trail}{}

\item{var}{}

\item{Altitude}{}

\item{necT}{}

\item{surfDecl}{}

\item{minR}{Minimum ROS (m/h)}

\item{targSp}{}

\item{testN}{}
}
\value{
list
}
\description{
Find likelihood and consequence of death for canopy trees
}
