% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{frameRisk}
\alias{frameRisk}
\title{Find likelihood and consequence of death for canopy trees}
\usage{
frameRisk(
  Flora,
  Structure,
  default.species.params,
  a = a,
  weather,
  lightning = 0.05,
  fires = 5,
  smoulder = 24,
  Extinction = 0.15,
  hourStep = 3,
  slopeM = 6.7,
  slopeSD = 4.9,
  slopeLength = 391,
  rangeDir = 270,
  mRiver = 0.38,
  lRiver = 0.06,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  Test = 80,
  sensitive = 1,
  girdleH = c(),
  woodDensity = 1000,
  phloem = 0.01,
  moisture = 0.6,
  vAir500 = 2,
  barkDensity = 300,
  bark = 0.04,
  comBark = 700,
  resBark = 0,
  bMoisture = 0.5,
  distance = 0.1,
  trail = 10,
  var = 10,
  Altitude = 200,
  necT = 60,
  surfDecl = 10,
  minR = 1,
  targSp = "a",
  testN = 5
)
}
\arguments{
\item{Flora}{}

\item{Structure}{}

\item{default.species.params}{}

\item{a}{}

\item{weather}{}

\item{lightning}{}

\item{fires}{Number of fires to model for each age}

\item{smoulder}{}

\item{Extinction}{}

\item{hourStep}{}

\item{slopeM}{}

\item{slopeSD}{}

\item{slopeLength}{}

\item{rangeDir}{}

\item{mRiver}{}

\item{lRiver}{}

\item{l}{}

\item{Ms}{}

\item{Pm}{}

\item{Mr}{}

\item{Test}{}

\item{sensitive}{}

\item{girdleH}{}

\item{woodDensity}{}

\item{phloem}{}

\item{moisture}{}

\item{vAir500}{}

\item{barkDensity}{}

\item{bark}{}

\item{comBark}{}

\item{resBark}{}

\item{bMoisture}{}

\item{distance}{}

\item{trail}{}

\item{var}{}

\item{Altitude}{}

\item{necT}{}

\item{surfDecl}{}

\item{minR}{Minimum ROS (m/h)}

\item{targSp}{}

\item{testN}{}
}
\value{
list
}
\description{
Find likelihood and consequence of death for canopy trees
}
