% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{frameSpread}
\alias{frameSpread}
\title{Finds distance of fire spread for a part of a landscape}
\usage{
frameSpread(
  base.params,
  weather,
  a,
  igLoc = -1,
  t = 1,
  hourStep = 3,
  strikeTime = 1,
  Extinction = 0.15,
  fReach = 0,
  fEdge = 1,
  fireArea = 100,
  wHeading = 0,
  firelineW = 1,
  slopeM = 6.7,
  slopeSD = 4.9,
  slopeLength = 391,
  mRiver = 0.38,
  lRiver = 0.06,
  Test = 80,
  sensitive = 1,
  girdleH,
  woodDensity = 1000,
  barkDensity = 300,
  bark = 0.04,
  comBark = 700,
  resBark = 0,
  phloem = 0.01,
  moisture = 0.6,
  bMoisture = 0.5,
  FPC = 0.5,
  distance = 2,
  trail = 1000,
  var = 10,
  Altitude = 200,
  vAir500 = 2,
  necT = 60,
  surfDecl = 10,
  minR = 1,
  targSp = "c",
  testN = 5,
  Strata,
  Species,
  Flora,
  Structure,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5
)
}
\arguments{
\item{base.params}{Parameter file}

\item{weather}{Formatted weather dataset}

\item{a}{}

\item{igLoc}{Ignition location between -1 (base of sunny slope), 0 (ridgeline) & 1 (base of shade slope)}

\item{t}{Record number in the weather set for which tio model}

\item{hourStep}{Number of hours for which to model}

\item{strikeTime}{}

\item{Extinction}{Extinction litter moisture content (Percent ODW)}

\item{fReach}{Initial flame reach (m)}

\item{fEdge}{Fire edge: Head = 1, Flank = 0, Tail = -1}

\item{fireArea}{}

\item{wHeading}{Wind direction in relation to slope exposure: Up sun slope = 1, cross slope = 0, down sun slope = -1}

\item{firelineW}{}

\item{slopeM}{Mean slope}

\item{slopeSD}{Slope standard deviation}

\item{slopeLength}{Distance from ridgeline to river}

\item{mRiver}{Likelihood of encountering a medium river}

\item{lRiver}{Likelihood of encountering a large river}

\item{Test}{}

\item{sensitive}{}

\item{girdleH}{}

\item{woodDensity}{}

\item{barkDensity}{}

\item{bark}{}

\item{comBark}{}

\item{resBark}{}

\item{phloem}{}

\item{moisture}{}

\item{bMoisture}{}

\item{FPC}{}

\item{distance}{}

\item{trail}{}

\item{var}{}

\item{Altitude}{}

\item{vAir500}{}

\item{necT}{}

\item{surfDecl}{}

\item{minR}{}

\item{targSp}{}

\item{testN}{}

\item{Strata}{}

\item{Species}{}

\item{Flora}{}

\item{Structure}{}

\item{l}{}

\item{Ms}{}

\item{Pm}{}

\item{Mr}{}
}
\value{
dataframe
}
\description{
Finds distance of fire spread for a part of a landscape
}
