% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{frameStratify}
\alias{frameStratify}
\title{frameStratify}
\usage{
frameStratify(veg, mStrat = 4, sepSig = 0.001)
}
\arguments{
\item{veg}{A dataframe listing plant species with columns describing crown dimensions using standardised names:
veg, pN, spName, base, top, he, ht, wid, Site, sN}

\item{mStrat}{Maximum number of strata}

\item{sepSig}{p value to define significant stratum separation}
}
\value{
Dataframe
}
\description{
Arranges survey data into strata using k-means clustering

Data are stratified into 2-4 strata, then the largest number of strata 
are chosen where p<0.01. If none qualify, then the most significant
division is chosen.

Strata are sorted by top height and appended to the original data
}
