% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{frameStratify}
\alias{frameStratify}
\title{Arranges survey data into strata using k-means clustering}
\usage{
frameStratify(
  veg,
  pN = "Point",
  spName = "Species",
  base = "base",
  top = "top",
  he = "he",
  ht = "ht",
  mStrat = 4
)
}
\arguments{
\item{veg}{A dataframe listing plant species with columns describing crown dimensions}

\item{pN}{The number of the point in the transect}

\item{spName}{Name of the field with the species name}

\item{base}{Name of the field with the base height}

\item{top}{Name of the field with the top height}

\item{he}{Name of the field with dimension he}

\item{ht}{Name of the field with dimension ht}
}
\value{
Dataframe
}
\description{
Data are stratified into 2-4 strata, then the largest number of strata 
are chosen where p<0.01. If none qualify, then the most significant
division is chosen.
}
\details{
Strata are sorted by top height and appended to the original data
}
