% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{frameSummaryBeta}
\alias{frameSummaryBeta}
\title{Summary table of fire behaviour, beta version}
\usage{
frameSummaryBeta(flames, sites, ros, surface, IP)
}
\arguments{
\item{flames}{The dataframe $FlameSummaries}

\item{sites}{The dataframe $Sites}

\item{ros}{The dataframe $ROS}

\item{surface}{The dataframe $SurfaceResults}

\item{IP}{The dataframe $IgnitionPaths}
}
\value{
dataframe
}
\description{
Summarises FRaME generated fire behaviour by RepId
}
