% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{frameSurvey}
\alias{frameSurvey}
\title{Processes field survey data into tables formatted for fire modelling}
\usage{
frameSurvey(
  dat,
  default.species.params,
  pN = "Point",
  spName = "Species",
  base = "base",
  top = "top",
  he = "he",
  ht = "ht",
  wid = "width",
  rec = "Site",
  sN = "SiteName",
  negEx = 1,
  max = 54.22,
  rate = 0.026,
  a = 3.35,
  b = 0.832,
  age = NA,
  surf = 10,
  density = 300,
  nsH = 0.5,
  cover = 0.8,
  aQ = NA,
  bQ = NA,
  cQ = NA,
  maxNS = NA,
  rateNS = NA
)
}
\arguments{
\item{dat}{The dataframe containing the formatted field survey data}

\item{default.species.params}{Plant traits database}

\item{pN}{The number of the point in the transect}

\item{spName}{Name of the field with the species name}

\item{base}{Name of the field with the base height}

\item{top}{Name of the field with the top height}

\item{he}{Name of the field with dimension he}

\item{ht}{Name of the field with dimension ht}

\item{wid}{Name of the field with dimension width}

\item{rec}{Name of the field with the record number}

\item{sN}{Optional field with a site name}

\item{max}{Maximum weight of surface litter (t/ha)}

\item{rate}{Growth rate for surface litter in a negative exponential function}

\item{age}{Optional field with site age}

\item{surf}{Weight of surface litter in t/ha}

\item{density}{Wood density (kg.m-3)}

\item{nsH}{Top height of suspended layer}

\item{cover}{Percent cover of suspended layer}

\item{aQ}{Parameter for a quadratic trend; leave as NA if trend is negative exponential}

\item{bQ}{Parameter for a quadratic trend; leave as NA if trend is negative exponential}

\item{cQ}{Parameter for a quadratic trend; leave as NA if trend is negative exponential}

\item{maxNS}{Parameter for a negative exponential trend; leave as NA if trend is quadratic}

\item{rateNS}{Parameter for a negative exponential trend; leave as NA if trend is quadratic}
}
\value{
list
}
\description{
Processes field survey data into tables formatted for fire modelling
}
