% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{frameSurvey}
\alias{frameSurvey}
\title{Processes field survey data into tables formatted for fire modelling}
\usage{
frameSurvey(
  dat,
  pN = "Point",
  spName = "Species",
  pBase = "base",
  pTop = "top",
  hE = "he",
  hT = "ht",
  wid = "width",
  rec = "Site",
  sN = "SiteName",
  max = 54,
  rate = 0.026,
  age = NA,
  surf = 10
)
}
\arguments{
\item{dat}{The dataframe containing the formatted field survey data}

\item{pN}{The number of the point in the transect}

\item{spName}{Name of the field with the species name}

\item{pBase}{Name of the field with the base height}

\item{pTop}{Name of the field with the top height}

\item{hE}{Name of the field with dimension he}

\item{hT}{Name of the field with dimension ht}

\item{wid}{Name of the field with dimension width}

\item{rec}{Name of the field with the record number}

\item{sN}{Optional field with a site name}

\item{max}{Maximum weight of surface litter (t/ha)}

\item{rate}{Growth rate for surface litter in a negative exponential function}

\item{age}{Optional field with site age}

\item{surf}{Weight of surface litter in t/ha}
}
\value{
list
}
\description{
Processes field survey data into tables formatted for fire modelling
}
