% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{frameTables}
\alias{frameTables}
\title{Constructs parameter files for FRaME using 
outputs from plant}
\usage{
frameTables(
  dat,
  tr,
  age,
  rec = 1,
  propSamp = 0.5,
  transects = 10,
  propDead = 0,
  leafForm = "Flat",
  lwRat = 3,
  leafA = 0.002547,
  ram = 5,
  ignitionTemp = 260,
  moist = 1,
  G.C_rat = 3,
  C.C_rat = 0.1,
  deltaL = 0.46,
  hw = 0,
  sLitter = 15,
  diameter = 0.005
)
}
\arguments{
\item{dat}{The results of run_scm_collect}

\item{tr}{Table of input traits}

\item{age}{Years since disturbance}

\item{rec}{Number of the record}

\item{propSamp}{Proportion of cohorts to test (0-1)}

\item{transects}{Number of repeats for each sample}

\item{propDead}{Proportion of foliage dead}

\item{leafForm}{Flat or Round}

\item{lwRat}{}

\item{leafA}{Area of a leaf in m2}

\item{ram}{}

\item{ignitionTemp}{Temperature of the endotherm (degC)}

\item{moist}{Leaf moisture (ratio moisture weight to dry weight)}

\item{G.C_rat}{Ratio of gaps to clumps of leaves}

\item{C.C_rat}{Ratio of clump to canopy size}

\item{deltaL}{Leaf density (g/cm3)}

\item{hw}{}

\item{sLitter}{Weight of surface litter (t/ha)}

\item{diameter}{Mean diameter of surface litter pieces (m)}
}
\description{
Constructs parameter files for FRaME using 
outputs from plant
}
