% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_interface.R
\name{frameTables}
\alias{frameTables}
\title{Constructs parameter files for FRaME using 
outputs from plant}
\usage{
frameTables(
  dat,
  age,
  rec = 1,
  sample = 0.5,
  transects = 10,
  propDead = 0,
  leafForm = "Flat",
  lwRat = 3,
  leafA = 0.002547,
  ram = 5,
  ignitionTemp = 260,
  moist = 1,
  G.C_rat = 3,
  C.C_rat = 0.1,
  density = 0.46,
  lat = -35,
  map = 1000,
  mat = 20,
  lma
)
}
\arguments{
\item{dat}{The results of run_scm_collect}

\item{age}{Years since disturbance}

\item{rec}{Number of the record}

\item{sample}{Proportion of cohorts to test (0-1)}

\item{transects}{Number of repeats for each sample}

\item{propDead}{Proportion of foliage dead}

\item{leafForm}{Flat or Round}

\item{leafA}{Area of a leaf in m2}

\item{ignitionTemp}{Temperature of the endotherm (degC)}

\item{moist}{Leaf moisture (ratio moisture weight to dry weight)}

\item{G.C_rat}{Ratio of gaps to clumps of leaves}

\item{C.C_rat}{Ratio of clump to canopy size}

\item{density}{Leaf density (g/cm3)}

\item{lat}{Latitude (degrees)}

\item{map}{Mean annual precipitation (mm)}

\item{mat}{Mean annual temperature (degC)}

\item{lma}{List of LMA values per species (kgm−2)}

\item{openness}{Ratio of gaps to clumps of leaves}

\item{clump}{Ratio of clump frontal area diameter to crown frontal area diameter}

\item{litter}{Weight of surface litter (t/ha)}

\item{diameter}{Mean diameter of surface litter pieces (m)}

\item{lwRatio}{Ratio of leaf length to width}
}
\description{
Constructs parameter files for FRaME using 
outputs from plant
}
