% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMC.R
\name{frameWeather}
\alias{frameWeather}
\title{Formats hourly weather data for frame, with DFMC}
\usage{
frameWeather(
  clim,
  m = 0.15,
  LAI = 3,
  WRF = 3,
  hCan = 20,
  rholitter = 550,
  litterW = 10,
  lat = -31.89,
  altitude = 8,
  slope = 6.7,
  slopeSD = 4.9,
  rangeDir = 270,
  dt = 3600,
  cardinal = FALSE,
  nelsonA = 5.2,
  nelsonB = -19,
  conLitter = 3e-04,
  sigma = 3000,
  EPS = 0.01
)
}
\arguments{
\item{clim}{A dataset with the fields:
tAM (9am temp, degC)
tPM (3pm temp, degC)
tMin (Minimum daily temp, degC)
tMax (Maximum daily temp, degC)
rhAM (9am Relative humidity, %)
rhPM (3pm Relative humidity, %)
wAM (9am wind, km/h)
wPM (3pm wind, km/h)
cAM (Morning cloud, oktas)
cPM (Afternoon cloud, oktas)
MSLP (Mean Sea Level Pressure, hPa)
Rain (Daily rainfall, mm)}

\item{m}{The starting moisture (proportion ODW)}

\item{LAI}{Leaf area index of the vegetation}

\item{WRF}{Wind reduction factor}

\item{hCan}{Height of the tree canopy (m)}

\item{rholitter}{Density of litter particles (kg/m3)}

\item{litterW}{Weight of litter (t/ha)}

\item{lat}{Latitude (degrees)}

\item{altitude}{(m)}

\item{slope}{(degrees)}

\item{slopeSD}{Standard deviation of the slope}

\item{rangeDir}{Cardinal direction of the ridgelines - either Nth/Sth (0) or west/east (270)}

\item{dt}{Seconds per time step}

\item{cardinal}{Simplify compass directions to cardinal (TRUE/FALSE)}

\item{nelsonA}{Constant from Nelson FMC model}

\item{nelsonB}{Constant from Nelson FMC model}

\item{conLitter}{Litter conductance (m/s)}

\item{sigma}{Surface area: volume ratio of litter particles (m2/m3)}

\item{EPS}{Sensitivity of iteration}
}
\description{
Function uses the 'Single differential equation model' of Matthews et al (2010)
}
