% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMC.R
\name{frameWeather}
\alias{frameWeather}
\title{Formats hourly weather data for frame, with DFMC}
\usage{
frameWeather(
  clim,
  m = 15,
  LAI = 3,
  WRF = 3,
  rholitter = 550,
  litterW = 10,
  lat = -31.89,
  altitude = 8,
  slope = 0,
  rangeDir = 0,
  dt = 3600,
  nelsonA = 5.2,
  nelsonB = -19,
  conLitter = 3e-04,
  sigma = 3000,
  EPS = 0.01
)
}
\arguments{
\item{clim}{A dataset with the fields:
tAM (9am temp, degC)
tPM (3pm temp, degC)
tMin (Minimum daily temp, degC)
tMax (Maximum daily temp, degC)
rhAM (9am Relative humidity, %)
rhPM (3pm Relative humidity, %)
wAM (9am wind, km/h)
wPM (3pm wind, km/h)
cAM (Morning cloud, oktas)
cPM (Afternoon cloud, oktas)
MSLP (Mean Sea Level Pressure, hPa)
Rain (Daily rainfall, mm)}

\item{m}{The starting moisture (proportion ODW)}

\item{dt}{Seconds per time step}

\item{nelsonA}{Constant from Nelson FMC model}

\item{nelsonB}{Constant from Nelson FMC model}

\item{conLitter}{Litter conductance (m/s)}

\item{sigma}{Surface area: volume ratio of litter particles (m2/m3)}

\item{EPS}{Sensitivity of iteration}

\item{tAir}{Atmospheric temperature (K)}

\item{vAir}{Wind speed (m/s)}

\item{pAir}{Atmospheric pressure (Pa)}

\item{rhAir}{Relative humidity (%)}

\item{rhoLitter}{Density of leaves (kg/m3)}

\item{insolation}{Solar energy at the soil surface (kW/m2)}
}
\description{
Function uses the 'Single differential equation model' of Matthews et al (2010)
}
