% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{htDyn}
\alias{htDyn}
\title{htDyn}
\usage{
htDyn(dat, thres = 5, pnts = 10, p = 0.05)
}
\arguments{
\item{dat}{The dataframe containing the input data,}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{pnts}{The number of points measured in a transect}

\item{p}{The maximum allowable p value for a model}
}
\value{
dataframe
}
\description{
Builds models for top-ht height allometry dynamics of surveyed Species

Input table requires the following fields:
Point - numbered point in a transect
Species - name of the surveyed Species
Age - age of the site since the triggering disturbance

Species that are less common than the set threshold are combined as "Minor Species"
}
