% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{ignitions}
\alias{ignitions}
\title{Finds randomised lightning ignition time and location}
\usage{
ignitions(weather, smoulder = 24, Extinction = 0.15)
}
\arguments{
\item{weather}{Hourly weather dataset from function frameWeather}

\item{smoulder}{Time for which an ignition may smoulder (hours)}

\item{Extinction}{Extinction moisture threshold}
}
\value{

}
\description{
Finds randomised lightning ignition time and location
}
