% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignitionTempPSFA.R
\name{ignitionTempPSFA}
\alias{ignitionTempPSFA}
\title{Calculates ignition temperature from proportion of silica free ash.}
\usage{
ignitionTempPSFA(prop)
}
\arguments{
\item{prop}{Value(s) in the interval (0, 1] for the proportion of 
silica free ash,}
}
\value{
Calculated ignition temperature(s) in degrees Celsius.
}
\description{
Calculates ignition temperature from proportion of silica free ash.
}

