% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{LAIcomm}
\alias{LAIcomm}
\title{Calculates LAI for a horizontal slice of a plant community}
\usage{
LAIcomm(base.params, yu = 100, yl = 0)
}
\arguments{
\item{base.params}{}

\item{yu}{Top of slice (m)}

\item{yl}{Base of slice (m)}
}
\value{
value
}
\description{
Calculates LAI for a horizontal slice of a plant community
}
