% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{litter}
\alias{litter}
\title{litter}
\usage{
litter(negEx = 1, max = 54.22, rate = 0.026, a = 3.35, b = 0.832, age = 10)
}
\arguments{
\item{negEx}{Value determining the model used. 
1 = olson, 2 = Burr}

\item{max}{Maximum weight (t/ha)}

\item{rate}{Growth rate for a negative exponential function}

\item{a}{Parameter in the Burr equation}

\item{b}{Parameter in the Burr equation}

\item{age}{The number of years since last fire}
}
\value{
value
}
\description{
Models the weight of surface litter from time since fire 
using either an Olson negative exponential function or a Burr curve
}
