% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{mammal}
\alias{mammal}
\title{Fire risk for an exposed arboreal mammal}
\usage{
mammal(
  Surf,
  Plant,
  Height = 1,
  distance = 5,
  trail = 360,
  diameter = 6,
  surfDecl = 10,
  var = 10,
  Pressure = 1013.25,
  Altitude = 0,
  RH = 0.51,
  bodyLength = 0.1,
  surfaceArea = 0.2,
  bodyMass = 1,
  protection = 0.0017,
  fibreCount = 100,
  fibreDiameter = 0.01,
  fibreCp = 2.5,
  fiberSolid = 0.5,
  skinCp = 3.5,
  skinDensity = 1020,
  skinK = 0.187,
  bodyTemp = 37,
  Shape = "Cylinder",
  updateProgress = NULL
)
}
\arguments{
\item{Surf}{The dataframe 'runs' exported from Monte Carlos as 'Summary.csv'}

\item{Plant}{The dataframe 'IP' exported from Monte Carlos as 'IP.csv'.}

\item{Height}{The height directly over ground (m) at which the species is expected to shelter from a fire.}

\item{distance}{The starting horizontal distance between the flame origin and the point (m)}

\item{trail}{Number of seconds to continue modelling after the fire has passed}

\item{diameter}{Diameter of the surface fuels burning (mm)}

\item{var}{The angle in degrees that the plume spreads above/below a central vector}

\item{Pressure}{Sea level atmospheric pressure (hPa)}

\item{Altitude}{Height above sea level (m)}

\item{RH}{Relative humidity (0-1)}

\item{bodyLength}{The "Characteristic length" of the animal (m)}

\item{surfaceArea}{The surface area of the animal (m^2)}

\item{bodyMass}{The mass of the animal (kg)}

\item{protection}{The thickness of fur covering the animal (m)}

\item{fibreCount}{The number of fibres per square mm}

\item{fibreDiameter}{The mean fibre diameter of hairs (mm)}

\item{fibreCp}{Specific heat of fibres (kJ/kg/C)}

\item{fiberSolid}{The proportion of the fibre (0-1)}

\item{skinCp}{Specific heat of the animal skin (kJ/kg/C)}

\item{skinDensity}{Density of the animal skin (kg/m3)}

\item{skinK}{Thermal conductivity of the animal skin (W/m/C)}

\item{bodyTemp}{The body temperature of the animal (deg C)}

\item{Shape}{The approximate shape of the animal - either "Flat", "Sphere", or "Cylinder"}

\item{updateProgress}{Progress bar for use in the dashboard}

\item{Specific_heat}{The specific heat of the fur (kJ/kg/deg C)}
}
\value{
dataframe
}
\description{
Calculates the degree of injury or likelihood of mortality 
to an exposed mammal caused by an approaching fire front
}
