% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threat.R
\name{phi}
\alias{phi}
\title{Configuration or view factor}
\usage{
phi(Fl, Fa, S, D, H)
}
\arguments{
\item{Fl}{}

\item{Fa}{}

\item{S}{}

\item{D}{}

\item{H}{}
}
\description{
Calculates the configuration factor for radiative heat transfer as used
in AS3959 for bushfire risk assessment of built structures
}
