% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{pHt}
\alias{pHt}
\title{Predicts plant Ht at a given age}
\usage{
pHt(mods, sp, Age = 10)
}
\arguments{
\item{mods}{A table of models fit to species, format as per modCollector}

\item{sp}{The name of a species in the table mods}

\item{Age}{The age of the plant (years since fire)}
}
\value{
value
}
\description{
Selects model from tabled models per species
}
