% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{plantFlame}
\alias{plantFlame}
\title{Collects flame segments for a specified plant in a stratum}
\usage{
plantFlame(paths, Stratum, Species, repId)
}
\arguments{
\item{paths}{Output table from the function repFlame}

\item{Stratum}{Name of the stratum being studied}

\item{Species}{Name of the species being studied}

\item{repId}{Number of the repId being studied}
}
\value{
dataframe
}
\description{
Collects flame segments for a specified plant in a stratum
}
\examples{
balga <- plantFlame(IPW, "NearSurface", "Xanthorrhoea preissii", 6)
}
