% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{plantRisk}
\alias{plantRisk}
\title{Calculates risk (likelihood & consequence) for plants in a given climate & terrain}
\usage{
plantRisk(
  dynDat,
  a,
  ignitions = 1,
  hourStep = 3,
  Area = 10,
  weather,
  vAir = 25,
  wStDev = 5,
  tAir = 30,
  dfmc = 0.05,
  slope = 0,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  Test = 80,
  fireN = 1
)
}
\arguments{
\item{dynDat}{}

\item{a}{}

\item{ignitions}{}

\item{hourStep}{}

\item{Area}{}

\item{weather}{}

\item{vAir}{}

\item{wStDev}{}

\item{tAir}{}

\item{dfmc}{}

\item{slope}{}

\item{l}{}

\item{Ms}{}

\item{Pm}{}

\item{Mr}{}

\item{Test}{}

\item{fireN}{}
}
\value{
list
}
\description{
Calculates risk (likelihood & consequence) for plants in a given climate & terrain
}
