% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{plantVarFrame}
\alias{plantVarFrame}
\title{Randomly modifies plant traits within defined ranges for non-deterministic predictions}
\usage{
plantVarFrame(
  base.params,
  Strata,
  Species,
  Flora,
  a,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.001
)
}
\arguments{
\item{base.params}{Parameter input table}

\item{Strata}{Strata descriptor table output by the function 'strata'}

\item{Species}{Species descriptor table output by the function 'species'}

\item{Flora}{Table output from the function buildFlora}

\item{a}{A unique identifier for the record being run}

\item{l}{Variation around input leaf dimensions}

\item{Ms}{Standard deviation of LFMC}

\item{Pm}{Multiplier of mean LFMC}

\item{Mr}{Truncates LFMC variability by +/- Mr * LFMC}
}
\value{
dataframe
}
\description{
Differs from plantVarS by using FRaME-updated tables
}
