% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{profileDet}
\alias{profileDet}
\title{Calculates a vertical wind profile}
\usage{
profileDet(base.params, slices = 10)
}
\arguments{
\item{base.params}{}

\item{slices}{Number of horizontal slices to use in calculation}
}
\value{
dataframe
}
\description{
Calculates a vertical wind profile
}
