% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{pseudoTransect}
\alias{pseudoTransect}
\title{Creates a dataset of plants with expected size and cover
randomly varied within natural ranges}
\usage{
pseudoTransect(
  Dynamics,
  pointRich,
  default.species.params,
  pnts = 10,
  Age = 10,
  growth = TRUE,
  thin = TRUE,
  prune = TRUE
)
}
\arguments{
\item{Dynamics}{Table of growth models as output by floraDynamics}

\item{pointRich}{table of species richness likelihood for a point
as output by function rich}

\item{default.species.params}{Plant traits database}

\item{pnts}{number of points in the transect}

\item{Age}{Age of the site in years}

\item{growth}{Logical - TRUE uses plant growth models in Dynamics,
otherwise mean values are used and plants are not grown}

\item{thin}{Logical - TRUE uses plant self-thinning models in Dynamics,
otherwise mean values are used and plant cover remains constant}

\item{prune}{Logical - TRUE uses plant self-pruning models in Dynamics,
otherwise mean values are used and plants are not self-pruned}
}
\value{
dataframe
}
\description{
Creates a dataset of plants with expected size and cover
randomly varied within natural ranges
}
