% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMC.R
\name{QSat}
\alias{QSat}
\title{saturation specific humidity}
\usage{
QSat(tAir = 293.46, pAir = 101180)
}
\arguments{
\item{tAir}{Temperature in Kelvin}

\item{pAir}{Atmospheric pressure in Pascals}
}
\description{
Uses Stull (1988) Eqn 7.5.2c and d
}
