% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLegacyParamFile.R
\name{readLegacyParamFile}
\alias{readLegacyParamFile}
\title{Reads a legacy format parameters file.}
\usage{
readLegacyParamFile(path, flat = FALSE)
}
\arguments{
\item{path}{full or relative path to the file}

\item{flat}{if \code{TRUE}, return parsed data as a single
data frame; if \code{FALSE}, return parsed data as a nested
list for strata, species and meta-data in a form suitable to
create Scala objects.}
}
\value{
Either a nested list or single data frame of parsed data.
}
\description{
Reads a legacy format parameters file.
}

