% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLegacyParamFile.R
\name{readLegacyParamFile}
\alias{readLegacyParamFile}
\title{Reads a legacy format parameters file.}
\usage{
readLegacyParamFile(path, raw = FALSE)
}
\arguments{
\item{path}{The full or relative path to the parameters file.}

\item{raw}{(logical) Whether to return raw parsed data.}
}
\value{
If \code{raw} is \code{TRUE}, a data frame of parsed data
  in raw form (useful for testing).
  If \code{raw} is \code{FALSE} (the default), a data frame in a form 
  suitable for use with \code{\link{ffm_run}}.
}
\description{
Reads a legacy format parameters file.
}
