% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{rich}
\alias{rich}
\title{rich}
\usage{
rich(dat, thres = 5, pnts = 10)
}
\arguments{
\item{dat}{The dataframe containing the input data,}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{pnts}{The number of points measured in a transect}
}
\value{
dataframe
}
\description{
Finds the distribution of species richness at a point

Input table requires the following fields:
Point - numbered point in a transect
Species - name of the surveyed Species
Age - age of the site since the triggering disturbance

Species that are less common than the set threshold are combined as "Minor Species"
}
