% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{richS}
\alias{richS}
\title{Finds the distribution of species richness per stratum}
\usage{
richS(dat, thres = 5, cols, pnts = 10, nstrat = 4)
}
\arguments{
\item{dat}{The dataframe containing the input data,}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{cols}{A list of the columns to be used for stratification, including base & top}

\item{pnts}{The number of points measured in a transect}

\item{nstrat}{The maximum number of strata}

\item{p}{The maximum allowable p value for a model}
}
\value{
dataframe
}
\description{
Input table requires the following fields:
Point - numbered point in a transect
Species - name of the surveyed Species
base - base height of each species
top - top height of each species
}
\details{
Species that are less common than the set threshold are combined as "Minor Species"
}
