% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{riskDynamics}
\alias{riskDynamics}
\title{Calculates forest risk parameters over a growth series
using parallel ports}
\usage{
riskDynamics(
  fireDat,
  tr,
  clim,
  lightning = 0.05,
  fires = 5,
  lat = -42.48,
  Altitude = 200,
  slopeM = 6.7,
  slopeSD = 4.9,
  slopeLength = 391,
  rangeDir = 270,
  mRiver = 0.38,
  lRiver = 0.06,
  cardinal = TRUE,
  smoulder = 24,
  Extinction = 0.15,
  m = 0.15,
  rholitter = 550,
  hourStep = 3,
  l = 0.1,
  Ms = 0.01,
  Pm = 1,
  Mr = 1.5,
  Test = 80,
  sensitive = 1,
  girdleH = c(),
  phloem = 0.01,
  moisture = 0.6,
  barkDensity = 300,
  bark = 0.04,
  comBark = 700,
  resBark = 0,
  bMoisture = 0.5,
  vAir500 = 2,
  testN = 5,
  distance = 2,
  trail = 1000,
  var = 10,
  necT = 60,
  surfDecl = 10,
  minR = 0.001,
  freeCores = 1
)
}
\arguments{
\item{fireDat}{List of three dataframes: Flora, Structure & default.species.params}

\item{tr}{}

\item{clim}{Formatted AM & PM weather conditions over a time sequence}

\item{lightning}{Number of lightning strikes per km2}

\item{fires}{Number of fires to model per age}

\item{lat}{Latitude (degrees)}

\item{Altitude}{Altitude (m.a.s.l.)}

\item{slopeM}{Mean slope (degrees)}

\item{slopeSD}{Standard deviation of slope (degrees)}

\item{slopeLength}{Distance from ridge to gully (m)}

\item{rangeDir}{}

\item{mRiver}{}

\item{lRiver}{}

\item{cardinal}{}

\item{smoulder}{}

\item{Extinction}{}

\item{m}{}

\item{rholitter}{}

\item{hourStep}{}

\item{l}{}

\item{Ms}{}

\item{Pm}{}

\item{Mr}{}

\item{Test}{}

\item{sensitive}{}

\item{girdleH}{}

\item{phloem}{}

\item{moisture}{}

\item{barkDensity}{}

\item{bark}{}

\item{comBark}{}

\item{resBark}{}

\item{bMoisture}{}

\item{vAir500}{}

\item{testN}{}

\item{distance}{}

\item{trail}{}

\item{var}{}

\item{necT}{}

\item{surfDecl}{}

\item{minR}{}

\item{freeCores}{}

\item{woodDensity}{}

\item{targSp}{}
}
\value{
list
}
\description{
Calculates forest risk parameters over a growth series
using parallel ports
}
