% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{river}
\alias{river}
\title{Support function to randomise river width}
\usage{
river(lRiver, mRiver)
}
\arguments{
\item{lRiver}{Likelihood of encountering a large river}

\item{mRiver}{Likelihood of encountering a medium river}
}
\description{
Support function to randomise river width
}
