% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earth.R
\name{satSoil}
\alias{satSoil}
\title{Find soil saturation}
\usage{
satSoil(texture = "loam", moisture = 0.3)
}
\arguments{
\item{texture}{Soil texture}

\item{moisture}{Volumetric water content}
}
\description{
Finds saturation from ODW moisture and texture

Field capacity of soils taken from Salter, P. J. & Williams, J. B.
The influence of texture on the moisture characteristics of soil.
V. Relationships between particle-size composition and moisturecontents
at the upper and lower limits of available-water. J. Soil Sci. 20, 126–131 (1969).
}
