% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMC.R
\name{simplefmc}
\alias{simplefmc}
\title{Finds the moisture content of dead leaf litter at a given time-step}
\usage{
simplefmc(
  m,
  tAir = 293.46,
  vAir = 3.52,
  pAir = 101180,
  rhAir = 58,
  dt = 3600,
  insolation = 150,
  rhoLitter = 550,
  nelsonA = 5.2,
  nelsonB = -19,
  conLitter = 3e-04,
  sigma = 3000,
  EPS = 0.01
)
}
\arguments{
\item{m}{The starting moisture (proportion ODW)}

\item{tAir}{Atmospheric temperature (K)}

\item{pAir}{Atmospheric pressure (Pa)}

\item{rhAir}{Relative humidity (%)}

\item{dt}{Seconds per time step}

\item{insolation}{Solar energy at the soil surface (kW/m2)}

\item{rhoLitter}{Density of leaves (kg/m3)}

\item{nelsonA}{Constant from Nelson FMC model}

\item{nelsonB}{Constant from Nelson FMC model}

\item{conLitter}{Litter conductance (m/s)}

\item{sigma}{Surface area: volume ratio of litter particles (m2/m3)}

\item{EPS}{Sensitivity of iteration}

\item{vAirWind}{speed (m/s)}
}
\description{
Function uses the 'Single differential equation model' of Matthews et al (2010)
}
