% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earth.R
\name{soil}
\alias{soil}
\title{Soil}
\usage{
soil(
  Surf,
  Plant,
  step = 0.01,
  diameter = 6,
  surface = 677,
  RH = 0.2,
  moisture = 0.1,
  distance = 50,
  trail = 600,
  var = 10,
  Pressure = 1013.25,
  Altitude = 0,
  texture = "sand",
  peat = 0,
  grain = "fine",
  unfrozen = 1,
  startTemp = 25,
  updateProgress = NULL
)
}
\arguments{
\item{Surf}{The dataframe 'runs' exported from Monte Carlos as 'Summary.csv'}

\item{Plant}{The dataframe 'IP' exported from Monte Carlos as 'IP.csv'.}

\item{step}{The increment of soil depth at which each calculation will be modelled (m)}

\item{diameter}{Diameter of the surface fuels burning (mm)}

\item{surface}{Temperature at the surface of the soil, under burning fuels}

\item{RH}{The relative humidity (0-1)}

\item{moisture}{The proportion oven-dry weight of moisture in the bark and wood}

\item{distance}{The furthest horizontal distance between the flame origin and the point (m)}

\item{trail}{Number of seconds to continue modelling after the front has passed}

\item{var}{The angle in degrees that the plume spreads above/below a central vector;defaults to 10}

\item{Pressure}{Sea level atmospheric pressure (hPa)}

\item{Altitude}{Height above sea level (m)}

\item{texture}{Soil texture. Allowable values are: "sand", "loamy sand", "sandy loam", "sandy clay loam",
"sand clay", "loam", "clay loam", "silt loam", "clay", "silty clay", "silty clay loam", "silt"}

\item{peat}{Organic proportion of the soil}

\item{grain}{Allowable values are "fine" or "coarse"}

\item{unfrozen}{Proportion of soil unfrozen, between 0 and 1}

\item{startTemp}{The initial temperature of the soil (C)}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\value{
dataframe
}
\description{
Soil heating

Calculates the dynamic heating of a soil at 1cm increments, to 5cm depth

Assumes all to be A horizon, with constant, uncompacted density

Utilises the output tables from 'threat' and adds to these
the Reynolds Number, heat transfer coefficients, Newton's convective energy transfer coefficient,
and the temperature of the object each second.

Reynolds Number utilises a standard formulation (e.g. Gordon, N. T., McMahon, T. A. & Finlayson, B. L.
Stream hydrology: an introduction for ecologists. (Wiley, 1992))

Convective heat transfer coefficients use the widely adopted formulations of
Williams, F. A. Urban and wildland fire phenomenology. Prog. Energy Combust. Sci. 8, 317–354 (1982),
and Drysdale, D. An introduction to fire dynamics. (John Wiley and Sons, 1985)
utilising a Prandtl number of 0.7.

Heat is transferred into the earth using Fourier's Law. Spread continues for a period after the
passage of the fire front, equal to the duration of the surface flame, as determined using
Burrows, N. D. Flame residence times and rates of weight loss of eucalypt forest fuel particles.
Int. J. Wildl. Fire 10, 137–143 (2001).

Default temperature of the resident flame is the average of the surface maximums in
Cawson, J. G., Nyman, P., Smith, H. G., Lane, P. N. J. & Sheridan, G. J.
How soil temperatures during prescribed burning affect soil water repellency,
infiltration and erosion. Geoderma 278, 12–22 (2016).

Heating area is set to 1m2, flat, with a characteristic length of 1m

Broad germination and seed death temperatures are based on
Auld, T. D. & O’Connel, M. A. Predicting patterns of post‐fire germination in
35 eastern Australian Fabaceae. Aust. J. Ecol. 16, 53–70 (1991).

Predicts death of fine roots at 60C, but does not yet include a time component.
}
