% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{speciesBuilder}
\alias{speciesBuilder}
\title{Builds the dataframe species.values}
\usage{
speciesBuilder(Flora, site, a)
}
\arguments{
\item{Flora}{}

\item{site}{}

\item{a}{}
}
\value{
dataframe
}
\description{
Builds the dataframe species.values
}
