% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{spotFire}
\alias{spotFire}
\title{Calculates spotting distance}
\usage{
spotFire(flameHeight, slope, FPC, windExposure, vAir, vAir500 = 2, fireArea)
}
\arguments{
\item{flameHeight}{Flame height (m)}

\item{slope}{Degrees}

\item{FPC}{Foliage Projective Cover (0 - 100 percent)}

\item{windExposure}{<1 protected, >=1 Exposed}

\item{vAir}{Wind speed (km/h)}

\item{vAir500}{Multiplier of wind speed to estimate 500 hPA wind}

\item{fireArea}{Area of active fire (ha)}
}
\value{

}
\description{
Calculates spotting distance
}
