% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{strataBuilder}
\alias{strataBuilder}
\title{Builds the dataframe strata.meta}
\usage{
strataBuilder(Structure, Flora, a)
}
\arguments{
\item{Structure}{}

\item{Flora}{}

\item{a}{}
}
\value{
dataframe
}
\description{
Builds the dataframe strata.meta
}
