% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{stratify}
\alias{stratify}
\title{Arranges survey data into strata using k-means clustering
Strata are sorted by top height and appended to the original data}
\usage{
stratify(
  veg,
  pN = "Point",
  spName = "Species",
  pBase = "base",
  pTop = "top",
  nstrat = 4
)
}
\arguments{
\item{veg}{A dataframe listing plant species with columns describing crown dimensions}

\item{pN}{The number of the point in the transect}

\item{spName}{Name of the field with the species name}

\item{pBase}{Name of the field with the base height}

\item{pTop}{Name of the field with the top height}

\item{nstrat}{Number of strata. Defaults to 4}
}
\value{
Dataframe
}
\description{
Arranges survey data into strata using k-means clustering
Strata are sorted by top height and appended to the original data
}
