% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{stratify}
\alias{stratify}
\title{Arranges survey data into strata using k-means clustering}
\usage{
stratify(veg, cols, nstrat = 4)
}
\arguments{
\item{veg}{A dataframe listing plant species with columns describing crown dimensions}

\item{cols}{A list of the columns that will be used for classification, and the species name}

\item{nstrat}{Number of strata. Defaults to 4}
}
\value{
Dataframe
}
\description{
Arranges survey data into strata using k-means clustering
}
