% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{stratRich}
\alias{stratRich}
\title{Calculates the range, mean and sd of 
species richness in each plant stratum}
\usage{
stratRich(
  dat,
  thres = 5,
  pnts = 10,
  pN = "Point",
  spName = "Species",
  base = "base",
  top = "top",
  he = "he",
  ht = "ht"
)
}
\arguments{
\item{dat}{The dataframe containing the input data}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{pnts}{The number of points measured in a transect}

\item{nstrat}{Maximum number of strata}
}
\value{
dataframe
}
\description{
Input table requires the following fields:
Point - numbered point in a transect
A field with the base height of the plants
A field with the top height of the plants
}
\details{
Species that are less common than the set threshold are combined as "Minor Species"
}
