% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{stratSite}
\alias{stratSite}
\title{Divides site data into consecutively numbered strata with
base and top heights}
\usage{
stratSite(dat, thres = 0, sepSig = 0.001)
}
\arguments{
\item{dat}{A dataframe listing plant species with columns describing crown dimensions using standardised names:
veg, pN, spName, base, top, he, ht, wid, Site, sN}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{sepSig}{}
}
\value{
dataframe
}
\description{
Species that are less common than the set threshold are combined as "Minor Species"
}
