% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{stratSite}
\alias{stratSite}
\title{Divides site data into consecutively numbered strata with
base and top heights}
\usage{
stratSite(
  dat,
  thres = 0,
  pnts = 10,
  pN = "Point",
  spName = "Species",
  pBase = "base",
  pTop = "top",
  hE = "he",
  hT = "ht"
)
}
\arguments{
\item{dat}{The dataframe containing the input data}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{pnts}{The number of points measured in a transect}

\item{pN}{The number of the point in the transect}

\item{spName}{Name of the field with the species name}

\item{pBase}{Name of the field with the base height}

\item{pTop}{Name of the field with the top height}

\item{hE}{Name of the field with dimension he}

\item{hT}{Name of the field with dimension ht}
}
\value{
dataframe
}
\description{
Input table requires the following fields:
Point - numbered point in a transect
A field with the base height of the plants
A field with the top height of the plants
}
\details{
Species that are less common than the set threshold are combined as "Minor Species"
}
