% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{stratSite}
\alias{stratSite}
\title{Divides site data into consecutively numbered strata with
base and top heights}
\usage{
stratSite(dat, thres = 0, pnts = 10, base = "base", top = "top")
}
\arguments{
\item{dat}{The dataframe containing the input data}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{pnts}{The number of points measured in a transect}

\item{base}{Name of the field with the base height}

\item{top}{Name of the field with tghe top height}
}
\value{
dataframe
}
\description{
Input table requires the following fields:
Point - numbered point in a transect
A field with the base height of the plants
A field with the top height of the plants
}
\details{
Species that are less common than the set threshold are combined as "Minor Species"
}
