% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{susp}
\alias{susp}
\title{Finds the height of near-surface litter}
\usage{
susp(
  default.species.params,
  density = 300,
  cover = 0.8,
  age = 10,
  aQ = NA,
  bQ = NA,
  cQ = NA,
  maxNS = NA,
  rateNS = NA,
  dec = TRUE
)
}
\arguments{
\item{default.species.params}{Plant traits database}

\item{density}{Wood density (kg.m-3)}

\item{cover}{Percent cover of suspended layer}

\item{age}{Years since last fire}

\item{aQ}{Parameter for a quadratic trend; leave as NA if trend is negative exponential}

\item{bQ}{Parameter for a quadratic trend; leave as NA if trend is negative exponential}

\item{cQ}{Parameter for a quadratic trend; leave as NA if trend is negative exponential}

\item{maxNS}{Asymptote for negative exponential increase in NS}

\item{rateNS}{Rate for negative exponential increase in NS}

\item{dec}{Logical - TRUE allows for quadratic model to decline as vegetation thins,}

\item{wNS}{Width of NS patches (m)}
}
\value{
List
}
\description{
Finds the height of near-surface litter
}
