% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threat.R
\name{tau}
\alias{tau}
\title{Atmospheric attenuation}
\usage{
tau(D = 200, flameTemp = 1300, temperature = 288, rh = 0.51)
}
\arguments{
\item{D}{}

\item{flameTemp}{}

\item{temperature}{}

\item{rh}{}
}
\description{
Calculates the atmospheric transmissivity of radiation as used
in AS3959 for bushfire risk assessment of built structures. Equations from
Fuss, S P & Hamins, A
An estimate of the correction applied to radiant flame measurements due to
attenuation by atmospheric CO2and H2O.
Fire Saf. J. 37, 181–190 (2002)
}
