% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{transectLong}
\alias{transectLong}
\title{Combines multiple transects into one}
\usage{
transectLong(alldata, pN = "Point")
}
\arguments{
\item{alldata}{Raw survey data with one or more transects}

\item{pN}{The name of the column that will be the number of the measurement point}
}
\value{
dataframe
}
\description{
Combines multiple transects into one
}
