% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_dynamics.R
\name{wDyn}
\alias{wDyn}
\title{Builds models for top-w height allometry dynamics of surveyed Species}
\usage{
wDyn(dat, thres = 5, pnts = 10, p = 0.05, bTest = 10, maxiter = 1000)
}
\arguments{
\item{dat}{The dataframe containing the input data,}

\item{thres}{The minimum percent cover (0-100) of a Species that will be analysed}

\item{pnts}{The number of points measured in a transect}

\item{p}{The maximum allowable p value for a model}

\item{bTest}{Multiples of mean + mRSE for which Burr & quadratic models can predict 
beyond the observed mean + standard deviation}

\item{maxiter}{The maximum number of iterations for model fitting}
}
\value{
dataframe
}
\description{
Input table requires the following fields:
Point - numbered point in a transect
Species - name of the surveyed Species
Age - age of the site since the triggering disturbance
}
\details{
Species that are less common than the set threshold are combined as "Minor Species"
}
