% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{windReduction}
\alias{windReduction}
\title{Finds the Wind Reduction Factor for a param file}
\usage{
windReduction(base.params, test = 1.2)
}
\arguments{
\item{base.params}{}

\item{test}{Height at which the WRF is calculated (m)}
}
\value{
value
}
\description{
Finds the Wind Reduction Factor for a param file
}
